% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_gcae_experiment_results.R
\name{analyse_gcae_experiment_results}
\alias{analyse_gcae_experiment_results}
\title{Analyse the \code{gcae_experiment_results}}
\usage{
analyse_gcae_experiment_results(
  gcae_experiment_results,
  gcae_experiment_params
)
}
\arguments{
\item{gcae_experiment_results}{the results of a call
to \link{do_gcae_experiment},
as checked by \link{check_gcae_experiment_results}}

\item{gcae_experiment_params}{parameters to run a full
\code{GCAE} experimenr (i.e. call \link{do_gcae_experiment}),
as created by \link{create_gcae_experiment_params} and checked
by \link{check_gcae_experiment_params}}
}
\value{
the result of \link[gganimate]{animate}
}
\description{
Analyse the \code{gcae_experiment_results}.
Creates pictures
}
\author{
Richèl J.C. Bilderbeek
}
