% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_qt_prediction.R
\name{analyse_qt_prediction}
\alias{analyse_qt_prediction}
\title{Determine how well the quantitative trait prediction worked}
\usage{
analyse_qt_prediction(datadir, trainedmodeldir, png_filename, verbose = FALSE)
}
\arguments{
\item{datadir}{directory where sample data is stored.
If not absolute: assumed relative to \verb{GenoCAE/} directory.
A valid \code{datadir} ends with a slash (\code{/})
and does not start with a full stop (\code{.})}

\item{trainedmodeldir}{name of the folder to save the trained model to.}

\item{png_filename}{name of the \code{PNG} file to be produced}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link[plinkr]{check_verbose} to detect if this argument is valid.}
}
\value{
Nothing. Will create a PNG file.
}
\description{
Determine how well the quantitative trait prediction worked
}
\author{
Richèl J.C. Bilderbeek
}
