% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_qt_prediction_depersonalized.R
\name{analyse_qt_prediction_depersonalized}
\alias{analyse_qt_prediction_depersonalized}
\title{Determine how well the quantitative trait prediction worked}
\usage{
analyse_qt_prediction_depersonalized(datadir, trainedmodeldir, unique_id)
}
\arguments{
\item{datadir}{directory where sample data is stored.
If not absolute: assumed relative to \verb{GenoCAE/} directory.
A valid \code{datadir} ends with a slash (\code{/})
and does not start with a full stop (\code{.})}

\item{trainedmodeldir}{name of the folder to save the trained model to.}

\item{unique_id}{the unique identifier of that setting}
}
\value{
Nothing. Will create some files.
}
\description{
Determine how well the quantitative trait prediction worked
}
\author{
Richèl J.C. Bilderbeek
}
