% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_mse_from_identity_line.R
\name{calc_mse_from_identity_line}
\alias{calc_mse_from_identity_line}
\title{Calculate the mean squared error from the identity line.}
\usage{
calc_mse_from_identity_line(true_values, estimated_values)
}
\arguments{
\item{true_values}{the true values}

\item{estimated_values}{the estimated values}
}
\value{
the mean squared error from the identity line
}
\description{
Calculate the mean squared error from the identity line
(i.e. the line \code{y = x}).
If the estimated values are equal to the true values, the MSE is zero.
}
\author{
Richèl J.C. Bilderbeek
}
