% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_nmse_from_identity_line.R
\name{calc_nmse_from_identity_line}
\alias{calc_nmse_from_identity_line}
\title{Calculate the normalized mean squared error from the identity line.}
\usage{
calc_nmse_from_identity_line(true_values, estimated_values)
}
\arguments{
\item{true_values}{the true values}

\item{estimated_values}{the estimated values}
}
\value{
the mean squared error from the identity line
}
\description{
Calculate the normalized mean squared error from the identity line
(i.e. the line \code{y = x}).
If the estimated values are equal to the true values, the MSE is zero.
}
\examples{
true_values_degrees <- seq(1, 360)
estimated_values_degrees <- true_values_degrees + stats::runif(n = 360)
nmse_degrees <- calc_nmse_from_identity_line(
  true_values = true_values_degrees,
  estimated_values = estimated_values_degrees
)
nmse_degrees

# Convert degrees to radians
true_values_radians <- 2.0 * pi * true_values_degrees / 360
estimated_values_radians <- 2.0 * pi * estimated_values_degrees / 360
nmse_radians <- calc_nmse_from_identity_line(
  true_values = true_values_radians,
  estimated_values = estimated_values_radians
)
nmse_radians

# mse_degrees and mse_radians are the same, due to their normalized scales
}
\seealso{
use \link{calc_mse_from_identity_line} to calculate the
(un-normalized) mean square error from the identity line
}
\author{
Richèl J.C. Bilderbeek
}
