% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_nmse_from_phenotype_predictions.R
\name{calc_nmse_from_phenotype_predictions}
\alias{calc_nmse_from_phenotype_predictions}
\title{Calculate the normalized mean squared error through time
from the phenotype predictions}
\usage{
calc_nmse_from_phenotype_predictions(phenotype_predictions_table)
}
\arguments{
\item{phenotype_predictions_table}{a table of phenotype predictions,
with a true and estimated value for all individuals at a certain
epoch}
}
\value{
a table with the normalized mean squared error through time
}
\description{
Calculate the normalized mean squared error through time
(see \link{calc_nmse_from_identity_line})
from the phenotype predictions,
}
\examples{
calc_nmse_from_phenotype_predictions(
  phenotype_predictions_table = create_test_phenotype_predictions_table()
)
}
\seealso{
use \link{calc_nmse_from_identity_line} to calculate the
normalized mean square error from the identity line for a point in time
}
\author{
Richèl J.C. Bilderbeek
}
