% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_r_squared_from_phenotype_predictions.R
\name{calc_r_squared_from_phenotype_predictions}
\alias{calc_r_squared_from_phenotype_predictions}
\title{Calculate the \code{r_squared} through time
from the phenotype predictions}
\usage{
calc_r_squared_from_phenotype_predictions(phenotype_predictions_table)
}
\arguments{
\item{phenotype_predictions_table}{a table of phenotype predictions,
with a true and estimated value for all individuals at a certain
epoch}
}
\value{
a table with the \code{r_squared} through time
}
\description{
Calculate the \code{r_squared} through time
from the phenotype predictions,
}
\examples{
calc_r_squared_from_phenotype_predictions(
  phenotype_predictions_table = create_test_phenotype_predictions_table()
)
}
\author{
Richèl J.C. Bilderbeek
}
