% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_analyse_epochs.R
\name{check_analyse_epochs}
\alias{check_analyse_epochs}
\title{Check if the \code{analyse_epochs} is valid}
\usage{
check_analyse_epochs(analyse_epochs)
}
\value{
Nothing. Will \link{stop} if \code{analyse_epochs} is invalid.
}
\description{
Check if the \code{analyse_epochs} is valid
Will \link{stop} if not.
}
\examples{
check_analyse_epochs(c(1, 2))
}
\author{
Richèl J.C. Bilderbeek
}
