% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_csv_filename.R
\name{check_csv_filename}
\alias{check_csv_filename}
\title{Check if the \code{csv_filename} is valid}
\usage{
check_csv_filename(csv_filename)
}
\arguments{
\item{csv_filename}{name of the comma-seperated (\code{.csv}) file to be produced,
as checked by \link{check_csv_filename}}
}
\value{
Nothing. Will \link{stop} if \code{csv_filename} is invalid.
}
\description{
Check if the \code{csv_filename} is valid
Will \link{stop} if not.
}
\examples{
check_csv_filename("my.csv")
}
\author{
Richèl J.C. Bilderbeek
}
