% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{check_data}
\alias{check_data}
\title{Check if the \code{data} is valid}
\usage{
check_data(data)
}
\arguments{
\item{data}{file prefix, not including path, of the data files.
The data files must be in EIGENSTRAT
or PLINK binary (\code{.bed}, \code{.bim}, \code{.fam}) format)}
}
\value{
Nothing. Will \link{stop} if \code{data} is invalid.
}
\description{
Check if the \code{data} is valid
Will \link{stop} if not.
}
\examples{
check_data("HumanOrigins249_tiny")
}
\seealso{
to check the input data files, use \link{check_gcae_input_data}
}
\author{
Richèl J.C. Bilderbeek
}
