% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data_opts_id.R
\name{check_data_opts_id}
\alias{check_data_opts_id}
\title{Check if the \code{data_opts_id} is valid}
\usage{
check_data_opts_id(data_opts_id)
}
\arguments{
\item{data_opts_id}{data options id,
corresponding to a file data_opts/data_opts_id.json
The data options affect how data is input to the model.}
}
\value{
Nothing. Will \link{stop} if \code{data_opts_id} is invalid.
}
\description{
Check if the \code{data_opts_id} is valid
Will \link{stop} if not.
}
\examples{
check_data_opts_id("b_0_4")
}
\author{
Richèl J.C. Bilderbeek
}
