% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_datadir.R
\name{check_datadir}
\alias{check_datadir}
\title{Check if the \code{datadir} is valid}
\usage{
check_datadir(datadir)
}
\arguments{
\item{datadir}{directory where sample data is stored}
}
\value{
Nothing. Will \link{stop} if \code{datadir} is invalid.
}
\description{
Check if the \code{datadir} is valid
Will \link{stop} if not.
}
\author{
Richèl J.C. Bilderbeek
}
