% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_evaluate_results.R
\name{check_evaluate_results}
\alias{check_evaluate_results}
\title{Check if the \code{evaluate_results} are valid.}
\usage{
check_evaluate_results(evaluate_results)
}
\arguments{
\item{evaluate_results}{the results of a call to
\link{parse_evaluate_filenames},
as checked by \link{check_evaluate_results}}
}
\value{
nothing
}
\description{
Check if the \code{evaluate_results} are valid.
Will \link{stop} if not.
}
\examples{
evaluate_filenames <- c(
  get_gcaer_filename("f1_score_3.csv"),
  get_gcaer_filename("f1_score_5.csv"),
  get_gcaer_filename("f1_scores_pops_epoch_1.csv")
)
evaluate_results <- parse_evaluate_filenames(evaluate_filenames)
check_evaluate_results(evaluate_results)
}
\author{
Richèl J.C. Bilderbeek
}
