% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_file_extension.R
\name{check_file_extension}
\alias{check_file_extension}
\title{Check if the \code{file_extension} is valid}
\usage{
check_file_extension(file_extension)
}
\arguments{
\item{file_extension}{a file extension, for example \code{.R}
or \code{.csv}.}
}
\value{
Nothing. Will \link{stop} if \code{file_extension} is invalid.
}
\description{
Check if the \code{file_extension} is valid
Will \link{stop} if not.
}
\examples{
check_file_extension(".R")
check_file_extension(".csv")
check_file_extension(".png")
}
\author{
Richèl J.C. Bilderbeek
}
