% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_folder_name.R
\name{check_folder_name}
\alias{check_folder_name}
\title{Check if the \code{folder_name} is valid}
\usage{
check_folder_name(folder_name)
}
\value{
Nothing. Will \link{stop} if \code{folder_name} is invalid.
}
\description{
Check if the \code{folder_name} is valid
Will \link{stop} if not.
}
\examples{
check_folder_name("my.csv")
}
\author{
Richèl J.C. Bilderbeek
}
