% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_gcae_args.R
\name{check_gcae_args}
\alias{check_gcae_args}
\title{Check if the arguments to call GCAE are valid.}
\usage{
check_gcae_args(args)
}
\arguments{
\item{args}{arguments for the \code{GCAE} executable}
}
\value{
Nothing. Will \link{stop} if the arguments to call GCAE are valid.
}
\description{
Check if the arguments to call GCAE are valid.
Will \link{stop} if not.
}
\details{
This function is as lenient as possible.
}
\examples{
check_gcae_args("--help")
}
\author{
Richèl J.C. Bilderbeek
}
