% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_gcae_experiment_params.R
\name{check_gcae_experiment_params}
\alias{check_gcae_experiment_params}
\title{Check if the options to run GCAE are valid.}
\usage{
check_gcae_experiment_params(gcae_experiment_params)
}
\value{
the GCAE options,
as can be checked by \link{check_gcae_experiment_params}
}
\description{
Check if the options to run GCAE are valid.
Will \link{stop} if not.
}
\details{
Note that this function does not check if GCAE is actually present in
that folder, similar to \link{normalizePath}
with \code{mustWork} set to \link{FALSE}
}
\examples{
gcae_experiment_params <- create_gcae_experiment_params(
  gcae_options = create_gcae_options(),
  gcae_setup = create_test_gcae_setup(),
  analyse_epochs = seq(1, 2, 3)
)
check_gcae_experiment_params(gcae_experiment_params)
}
\author{
Richèl J.C. Bilderbeek
}
