% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_gcae_experiment_params_filename.R
\name{check_gcae_experiment_params_filename}
\alias{check_gcae_experiment_params_filename}
\title{Check if the \code{gcae_experiment_params_filename} is valid}
\usage{
check_gcae_experiment_params_filename(gcae_experiment_params_filename)
}
\arguments{
\item{gcae_experiment_params_filename}{name of the file to
save/read \code{gcae_experiment_params} to/from,
using \link{read_gcae_experiment_params_file} and
\link{save_gcae_experiment_params}}
}
\value{
Nothing.
Will \link{stop} if \code{gcae_experiment_params_filename} is invalid.
}
\description{
Check if the \code{gcae_experiment_params_filename} is valid
Will \link{stop} if not.
}
\examples{
check_gcae_experiment_params_filename("gcae_experiment_params.csv")
}
\author{
Richèl J.C. Bilderbeek
}
