% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_gcae_experiment_results.R
\name{check_gcae_experiment_results}
\alias{check_gcae_experiment_results}
\title{Check if the \code{gcae_experiment_results} is valid.}
\usage{
check_gcae_experiment_results(gcae_experiment_results)
}
\arguments{
\item{gcae_experiment_results}{the results of a call
to \link{do_gcae_experiment},
as checked by \link{check_gcae_experiment_results}}
}
\value{
nothing
}
\description{
Check if the \code{gcae_experiment_results} is valid.
Will \link{stop} if not.
}
\author{
Richèl J.C. Bilderbeek
}
