% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_gcae_folder.R
\name{check_gcae_folder}
\alias{check_gcae_folder}
\title{Check if the \code{gcae_folder} is valid}
\usage{
check_gcae_folder(gcae_folder)
}
\arguments{
\item{gcae_folder}{folder where \code{GCAE} is installed}
}
\value{
Nothing. Will \link{stop} if \code{gcae_folder} is invalid.
}
\description{
Check if the \code{gcae_folder} is valid
Will \link{stop} if not.
}
\examples{
check_gcae_folder("my_folder")
}
\author{
Richèl J.C. Bilderbeek
}
