% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_gcae_input_data.R
\name{check_gcae_input_data}
\alias{check_gcae_input_data}
\title{Check if the \code{gcae_input_data} is valid}
\usage{
check_gcae_input_data(gcae_input_data)
}
\value{
Nothing. Will \link{stop} if \code{gcae_input_data} is invalid.
}
\description{
Check if the \code{gcae_input_data} is valid
Will \link{stop} if not.
}
\examples{
gcae_input_filenames <- create_gcae_input_files_1()
gcae_input_data <- read_gcae_input_files(gcae_input_filenames)
check_gcae_input_data(gcae_input_data)

# Clean up
file.remove(as.character(unlist(gcae_input_filenames)))
}
\author{
Richèl J.C. Bilderbeek
}
