% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_gcae_input_filenames.R
\name{check_gcae_input_filenames}
\alias{check_gcae_input_filenames}
\title{Check if the GenoCAE input filenames \link{list} is valid.}
\usage{
check_gcae_input_filenames(gcae_input_filenames)
}
\arguments{
\item{gcae_input_filenames}{the GCAE input filenames,
in the form of a \link{list} with the following elements:
\itemize{
\item \code{bed_filename}: path to a \code{.bed} file,
as can be read by \link[plinkr]{read_plink_bed_file}
\item \code{bim_filename}: path to a \code{.bim} file,
as can be read by \link[plinkr]{read_plink_bim_file}
\item \code{fam_filename}: path to a \code{.fam} file,
as can be read by \link[plinkr]{read_plink_fam_file}
\item \code{phe_filename}: path to a \code{.phe} file,
as can be read by \link[plinkr]{read_plink_phe_file}
\item \code{labels_filename}: path to a \code{.csv} file
as can be read by \link{read_labels_file}
}}
}
\value{
Nothing.
Will \link{stop} if the GenoCAE input filenames \link{list} is invalid.
}
\description{
Check if the GenoCAE input filenames \link{list} is valid.
Will \link{stop} if not
}
\examples{
# Create the files
gcae_input_filenames <- create_gcae_input_files_1()
check_gcae_input_filenames(gcae_input_filenames)

# Clean up
file.remove(as.character(unlist(gcae_input_filenames)))
}
\author{
Richèl J.C. Bilderbeek
}
