% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_gcae_is_installed.R
\name{check_gcae_is_installed}
\alias{check_gcae_is_installed}
\title{Check if GCAE is installed}
\usage{
check_gcae_is_installed(gcae_options = create_gcae_options())
}
\arguments{
\item{gcae_options}{options to run GCAE,
as created by \link{create_gcae_options}}
}
\value{
Nothing. Will \link{stop} if GCAE is not installed.
}
\description{
Check if GCAE is installed. Will \link{stop} if not
}
\author{
Richèl J.C. Bilderbeek
}
