% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_gcae_options.R
\name{check_gcae_options}
\alias{check_gcae_options}
\title{Check if the options to run GCAE are valid.}
\usage{
check_gcae_options(gcae_options)
}
\arguments{
\item{gcae_options}{options to run GCAE,
as created by \link{create_gcae_options}}
}
\value{
the GCAE options,
as can be checked by \link{check_gcae_options}
}
\description{
Check if the options to run GCAE are valid.
Will \link{stop} if not.
}
\details{
Note that this function does not check if GCAE is actually present in
that folder, similar to \link{normalizePath}
with \code{mustWork} set to \link{FALSE}
}
\examples{
check_gcae_options(create_gcae_options())
}
\author{
Richèl J.C. Bilderbeek
}
