% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_gcae_repo_is_cloned.R
\name{check_gcae_repo_is_cloned}
\alias{check_gcae_repo_is_cloned}
\title{Check if the GCAE repo is cloned. Will \link{stop} if not.}
\usage{
check_gcae_repo_is_cloned(gcae_options = create_gcae_options())
}
\arguments{
\item{gcae_options}{options to run GCAE,
as created by \link{create_gcae_options}}
}
\value{
nothings
}
\description{
Check if the GCAE repo is cloned. Will \link{stop} if not.
}
\examples{
if (has_cloned_gcae_repo()) {
  check_gcae_repo_is_cloned()
}
}
\seealso{
use \link{clone_gcae_repo} to clone the GCAE repo
}
\author{
Richèl J.C. Bilderbeek
}
