% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_gcae_script_is_fixed.R
\name{check_gcae_script_is_fixed}
\alias{check_gcae_script_is_fixed}
\title{Check if the GCAE script is fixed}
\usage{
check_gcae_script_is_fixed(gcae_options = create_gcae_options())
}
\arguments{
\item{gcae_options}{options to run GCAE,
as created by \link{create_gcae_options}}
}
\value{
Nothing
}
\description{
Check if the GCAE script is fixed
Will \link{stop} if not
}
\author{
Richèl J.C. Bilderbeek
}
