% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_gcae_setup.R
\name{check_gcae_setup}
\alias{check_gcae_setup}
\title{Check if the options to run GCAE are valid.}
\usage{
check_gcae_setup(gcae_setup)
}
\arguments{
\item{gcae_setup}{GCAE model setup,
as created by \link{create_gcae_setup}}
}
\value{
the GCAE options,
as can be checked by \link{check_gcae_setup}
}
\description{
Check if the options to run GCAE are valid.
Will \link{stop} if not.
}
\details{
Note that this function does not check if GCAE is actually present in
that folder, similar to \link{normalizePath}
with \code{mustWork} set to \link{FALSE}
}
\examples{
gcae_setup <- create_test_gcae_setup()
check_gcae_setup(gcae_setup)
}
\author{
Richèl J.C. Bilderbeek
}
