% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_gcae_setup_filename.R
\name{check_gcae_setup_filename}
\alias{check_gcae_setup_filename}
\title{Check if the \code{gcae_setup_filename} is valid}
\usage{
check_gcae_setup_filename(gcae_setup_filename)
}
\arguments{
\item{gcae_setup_filename}{name of a file to save/load a \code{gcae_setup}
to/from}
}
\value{
Nothing. Will \link{stop} if \code{gcae_setup_filename} is invalid.
}
\description{
Check if the \code{gcae_setup_filename} is valid
Will \link{stop} if not.
}
\examples{
check_gcae_setup_filename("gcae_setup.csv")
}
\author{
Richèl J.C. Bilderbeek
}
