% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_gcae_train_files_are_present.R
\name{check_gcae_train_files_are_present}
\alias{check_gcae_train_files_are_present}
\title{Check if the files to run \link{gcae_train} are present.}
\usage{
check_gcae_train_files_are_present(gcae_setup)
}
\arguments{
\item{gcae_setup}{GCAE model setup,
as created by \link{create_gcae_setup},
read from file by \link{read_gcae_setup_file},
checked by \link{check_gcae_setup},
and saved to file by \link{save_gcae_setup}.}
}
\value{
nothing
}
\description{
Check if the files to run \link{gcae_train} are present.
Will \link{stop} if not.
}
\author{
Richèl J.C. Bilderbeek
}
