% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_gcae_version.R
\name{check_gcae_version}
\alias{check_gcae_version}
\title{Check if the GCAE version is valid, will stop if not}
\usage{
check_gcae_version(gcae_version)
}
\arguments{
\item{gcae_version}{version of GCAE, e.g. \code{"1.7"}
Use \link{get_gcae_version} to get the \code{GCAE} version.
Use \link{get_gcae_versions} to get all the supported \code{GCAE} versions.}
}
\description{
Check if the GCAE version is valid, will stop if not
}
\examples{
check_gcae_version("1.0")
}
\author{
Richèl J.C. Bilderbeek
}
