% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_genotype_concordances_table.R
\name{check_genotype_concordances_table}
\alias{check_genotype_concordances_table}
\title{Create a \code{genotype_concordances_table} to be used in testing}
\usage{
check_genotype_concordances_table(genotype_concordances_table)
}
\arguments{
\item{genotype_concordances_table}{genotype concordances
in time,
as can be checked by \link{check_genotype_concordances_table}}
}
\value{
a \code{genotype_concordances_table}
}
\description{
Create a \code{genotype_concordances_table} to be used in testing
}
\examples{
check_genotype_concordances_table(create_test_genotype_concordances_table())
}
\author{
Richèl J.C. Bilderbeek
}
