% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_gif_filename.R
\name{check_gif_filename}
\alias{check_gif_filename}
\title{Check if the \code{gif_filename} is valid}
\usage{
check_gif_filename(gif_filename)
}
\arguments{
\item{gif_filename}{name of a GIF file,
as checked by \link{check_gif_filename}}
}
\value{
Nothing. Will \link{stop} if \code{gif_filename} is invalid.
}
\description{
Check if the \code{gif_filename} is valid
Will \link{stop} if not.
}
\examples{
check_gif_filename("my.gif")
}
\author{
Richèl J.C. Bilderbeek
}
