% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input_files_are_present.R
\name{check_input_files_are_present}
\alias{check_input_files_are_present}
\title{Check if all the files needed for \link{do_gcae_experiment}
are present.}
\usage{
check_input_files_are_present(gcae_experiment_params)
}
\arguments{
\item{gcae_experiment_params}{parameters to run a full
\code{GCAE} experimenr (i.e. call \link{do_gcae_experiment}),
as created by \link{create_gcae_experiment_params} and checked
by \link{check_gcae_experiment_params}}
}
\value{
Nothing.
Will \link{stop} if any files needed for \link{do_gcae_experiment}
is absent.
}
\description{
Check if all the files needed for \link{do_gcae_experiment}
are present.
Will \link{stop} if not.
}
\author{
Richèl J.C. Bilderbeek
}
