% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_labels_table.R
\name{check_labels_table}
\alias{check_labels_table}
\title{Check if the \code{labels_table} is valid}
\usage{
check_labels_table(labels_table)
}
\arguments{
\item{labels_table}{a table containing super-
and -population labels, as can be read by \link{read_labels_table}.
Note that \code{GCAE} does not use column names, \code{gcaer} adds these
for convenience}
}
\value{
Nothing. Will \link{stop} if \code{labels_table} is invalid.
}
\description{
Check if the \code{labels_table} is valid
Will \link{stop} if not.
}
\details{
A \code{labels_table} is a \link[tibble]{tibble}
with two columns and zero, one or more rows.
The columns are named \code{population}
and \code{super_population}
and expresses which population (e.g. 'The Netherlands')
is part of which superpopulation (e.g. 'Europe').
There are as much rows as there are populations.
}
\author{
Richèl J.C. Bilderbeek
}
