% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_losses_from_train_t_table.R
\name{check_losses_from_train_t_table}
\alias{check_losses_from_train_t_table}
\title{Check if the \code{losses_from_train_t_table} is valid}
\usage{
check_losses_from_train_t_table(losses_from_train_t_table)
}
\arguments{
\item{losses_from_train_t_table}{losses from training, measured
in the training step,
as can be checked by \link{check_losses_from_train_t_table}}
}
\value{
Nothing.
Will \link{stop} if \code{losses_from_train_t_table} is invalid.
}
\description{
Check if the \code{losses_from_train_t_table} is valid
Will \link{stop} if not.
}
\examples{
losses_from_train_t_table <- create_test_losses_from_train_t_table()
check_losses_from_train_t_table(losses_from_train_t_table)
}
\author{
Richèl J.C. Bilderbeek
}
