% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_losses_from_train_v_table.R
\name{check_losses_from_train_v_table}
\alias{check_losses_from_train_v_table}
\title{Check if the \code{losses_from_train_v_table} is valid}
\usage{
check_losses_from_train_v_table(losses_from_train_v_table)
}
\value{
Nothing.
Will \link{stop} if \code{losses_from_train_v_table} is invalid.
}
\description{
Check if the \code{losses_from_train_v_table} is valid
Will \link{stop} if not.
}
\examples{
losses_from_train_v_table <- create_test_losses_from_train_v_table()
check_losses_from_train_v_table(losses_from_train_v_table)
}
\author{
Richèl J.C. Bilderbeek
}
