% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_metric.R
\name{check_metric}
\alias{check_metric}
\title{Check if the \code{metric} is valid}
\usage{
check_metric(metric)
}
\value{
Nothing. Will \link{stop} if \code{metric} is invalid.
}
\description{
Check if the \code{metric} is valid
Will \link{stop} if not.
}
\examples{
check_metric("hull_error")
check_metric("f1_score_3")
check_metric("f1_score_5")
}
\author{
Richèl J.C. Bilderbeek
}
