% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_metrics.R
\name{check_metrics}
\alias{check_metrics}
\title{Check if the \code{metrics} is valid}
\usage{
check_metrics(metrics)
}
\arguments{
\item{metrics}{the metrics, as checked by \link{check_metrics}}
}
\value{
Nothing. Will \link{stop} if \code{metrics} is invalid.
}
\description{
Check if the \code{metrics} is valid
Will \link{stop} if not.
}
\details{
\code{metrics} is a comma-separated list of \code{metric},
as checked by \link{check_metric}
}
\examples{
check_metrics("hull_error")
check_metrics("f1_score_3")
check_metrics("hull_error,f1_score_3")
}
\author{
Richèl J.C. Bilderbeek
}
