% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_model.R
\name{check_model}
\alias{check_model}
\title{Check if the \code{model} is valid}
\usage{
check_model(model)
}
\arguments{
\item{model}{the \code{GCAE} architecture,
as checked by \link{check_model}.
Use \link{read_model_file} to read a \code{GCAE} architecture from file.
Use \link{save_model} to save the \code{model} to a JSON file.
Use \link{plot_model} to plot the \code{model}.
Use \link{save_model_plot} to save a plot of the \code{model} to file.}
}
\value{
Nothing. Will \link{stop} if \code{model} is invalid.
}
\description{
Check if the \code{model} is valid
Will \link{stop} if not.
}
\examples{
# An example file
model_filename <- get_gcaer_filename("M0.json")
model <- read_model_file(model_filename)
check_model(model)

if (is_gcae_repo_cloned()) {
  # A GCAE file
  model_filename <- get_gcae_model_filename("M0")
  model <- read_model_file(model_filename)
  check_model(model)

  model_filename <- get_gcae_model_filename("M1")
  model <- read_model_file(model_filename)
  check_model(model)
}
}
\author{
Richèl J.C. Bilderbeek
}
