% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_model_id.R
\name{check_model_id}
\alias{check_model_id}
\title{Check if the \code{model_id} is valid}
\usage{
check_model_id(model_id)
}
\arguments{
\item{model_id}{model id, corresponding to a file models/model_id.json}
}
\value{
Nothing. Will \link{stop} if \code{model_id} is invalid.
}
\description{
Check if the \code{model_id} is valid
Will \link{stop} if not.
}
\examples{
check_model_id("M1")
}
\author{
Richèl J.C. Bilderbeek
}
