% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_nmse_in_time_table.R
\name{check_nmse_in_time_table}
\alias{check_nmse_in_time_table}
\title{Check if the \code{nmse_in_time_table} is valid}
\usage{
check_nmse_in_time_table(nmse_in_time_table)
}
\arguments{
\item{nmse_in_time_table}{normalized mean squared error in time,
as can be checked by \link{check_nmse_in_time_table}}
}
\value{
Nothing.
Will \link{stop} if \code{nmse_in_time_table} is invalid.
}
\description{
Check if the \code{nmse_in_time_table} is valid
Will \link{stop} if not.
}
\examples{
nmse_in_time_table <- create_test_nmse_in_time_table()
check_nmse_in_time_table(nmse_in_time_table)
}
\author{
Richèl J.C. Bilderbeek
}
