% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ormr_folder_name.R
\name{check_ormr_folder_name}
\alias{check_ormr_folder_name}
\title{Check if the \code{ormr_folder_name} is valid}
\usage{
check_ormr_folder_name(ormr_folder_name)
}
\arguments{
\item{ormr_folder_name}{folder where the link[ormr]{ormr} R package
installs the Python packages}
}
\value{
Nothing. Will \link{stop} if \code{ormr_folder_name} is invalid.
}
\description{
Check if the \code{ormr_folder_name} is valid
Will \link{stop} if not.
}
\examples{
check_ormr_folder_name("my_folder")
}
\author{
Richèl J.C. Bilderbeek
}
