% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pheno_model_filename.R
\name{check_pheno_model_filename}
\alias{check_pheno_model_filename}
\title{Check if the \code{pheno_model_filename} is valid}
\usage{
check_pheno_model_filename(pheno_model_filename)
}
\arguments{
\item{pheno_model_filename}{name of a file that stores
the phenotype prediction \code{GCAE} architecture,
as checked by \link{check_pheno_model_filename}.
Use \link{read_pheno_model_file} to read a \code{GCAE}
pehnotype prediction architecture from that file.}
}
\value{
Nothing. Will \link{stop} if \code{pheno_model_filename} is invalid.
}
\description{
Check if the \code{pheno_model_filename} is valid
Will \link{stop} if not.
}
\examples{
check_pheno_model_filename("p0.json")
check_pheno_model_filename("~/p0.json")
check_pheno_model_filename("/home/me/p2.json")
}
\seealso{
use \link{check_model_filename} to check if
an autoencoder model filename is correct
}
\author{
Richèl J.C. Bilderbeek
}
