% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pheno_model_id.R
\name{check_pheno_model_id}
\alias{check_pheno_model_id}
\title{Check if the \code{pheno_model_id} is valid}
\usage{
check_pheno_model_id(pheno_model_id)
}
\arguments{
\item{pheno_model_id}{phenotype model ID,
as checked by \link{check_pheno_model_id}}
}
\value{
Nothing. Will \link{stop} if \code{pheno_model_id} is invalid.
}
\description{
Check if the \code{pheno_model_id} is valid
Will \link{stop} if not.
}
\examples{
check_pheno_model_id("p0")
check_pheno_model_id("p1")
check_pheno_model_id("p2")
}
\seealso{
Use \link{check_model_id} to check if a \code{model_id}
is valid
}
\author{
Richèl J.C. Bilderbeek
}
