% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_phenotype_predictions_table.R
\name{check_phenotype_predictions_table}
\alias{check_phenotype_predictions_table}
\title{Check if the \code{phenotype_predictions_table} is valid}
\usage{
check_phenotype_predictions_table(phenotype_predictions_table)
}
\arguments{
\item{phenotype_predictions_table}{a table of phenotype predictions,
with a true and estimated value for all individuals at a certain
epoch}
}
\value{
Nothing.
Will \link{stop} if \code{phenotype_predictions_table} is invalid.
}
\description{
Check if the \code{phenotype_predictions_table} is valid
Will \link{stop} if not.
}
\examples{
phenotype_predictions_table <- create_test_phenotype_predictions_table()
check_phenotype_predictions_table(phenotype_predictions_table)
}
\author{
Richèl J.C. Bilderbeek
}
