% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pip_is_installed.R
\name{check_pip_is_installed}
\alias{check_pip_is_installed}
\title{Check if \code{pip} is installed}
\usage{
check_pip_is_installed(gcae_options = create_gcae_options())
}
\arguments{
\item{gcae_options}{options to run GCAE,
as created by \link{create_gcae_options}}
}
\value{
Nothing
}
\description{
Check if \code{pip} is installed.
Will \link{stop} if not
}
\examples{
if (is_pip_installed()) {
  check_pip_is_installed
}
}
\author{
Richèl J.C. Bilderbeek
}
