% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_png_filename.R
\name{check_png_filename}
\alias{check_png_filename}
\title{Check if the \code{png_filename} is valid}
\usage{
check_png_filename(png_filename)
}
\arguments{
\item{png_filename}{name of the \code{PNG} file to be produced,
as checked by \link{check_png_filename}}
}
\value{
Nothing. Will \link{stop} if \code{png_filename} is invalid.
}
\description{
Check if the \code{png_filename} is valid
Will \link{stop} if not.
}
\examples{
check_png_filename("my.png")
}
\author{
Richèl J.C. Bilderbeek
}
