% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_python_version.R
\name{check_python_version}
\alias{check_python_version}
\title{Check if the \code{python_version} is valid}
\usage{
check_python_version(python_version)
}
\arguments{
\item{python_version}{the Python version}
}
\value{
Nothing. Will \link{stop} if \code{python_version} is invalid.
}
\description{
Check if the \code{python_version} is valid
Will \link{stop} if not.
}
\examples{
check_python_version("3.6")
}
\author{
Richèl J.C. Bilderbeek
}
