% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_r_squared_in_time_table.R
\name{check_r_squared_in_time_table}
\alias{check_r_squared_in_time_table}
\title{Check if the \code{r_squared_in_time_table} is valid}
\usage{
check_r_squared_in_time_table(r_squared_in_time_table)
}
\value{
Nothing.
Will \link{stop} if \code{r_squared_in_time_table} is invalid.
}
\description{
Check if the \code{r_squared_in_time_table} is valid
Will \link{stop} if not.
}
\examples{
r_squared_in_time_table <- create_test_r_squared_in_time_table()
check_r_squared_in_time_table(r_squared_in_time_table)
}
\author{
Richèl J.C. Bilderbeek
}
