% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_resume_from.R
\name{check_resume_from}
\alias{check_resume_from}
\title{Check if the \code{resume_from} is valid}
\usage{
check_resume_from(resume_from)
}
\arguments{
\item{resume_from}{the epoch to resume from,
as checked by \link{chgeck_resume_from}}
}
\value{
Nothing. Will \link{stop} if \code{resume_from} is invalid.
}
\description{
Check if the \code{resume_from} is valid
Will \link{stop} if not.
}
\examples{
check_resume_from(100)
}
\author{
Richèl J.C. Bilderbeek
}
