% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_save_interval.R
\name{check_save_interval}
\alias{check_save_interval}
\title{Check if the \code{save_interval} is valid}
\usage{
check_save_interval(save_interval)
}
\arguments{
\item{save_interval}{epoch intervals at which to save state of model,
and at which to calculate the valid loss}
}
\value{
Nothing. Will \link{stop} if \code{save_interval} is invalid.
}
\description{
Check if the \code{save_interval} is valid
Will \link{stop} if not.
}
\author{
Richèl J.C. Bilderbeek
}
