% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_scores_per_pop_table.R
\name{check_scores_per_pop_table}
\alias{check_scores_per_pop_table}
\title{Create a \code{scores_per_pop_table} to be used in testing}
\usage{
check_scores_per_pop_table(scores_per_pop_table)
}
\arguments{
\item{scores_per_pop_table}{table that contains the dimensionality
reduction scores per population}
}
\value{
a \code{scores_per_pop_table}
}
\description{
Create a \code{scores_per_pop_table} to be used in testing
}
\examples{
check_scores_per_pop_table(create_test_scores_per_pop_table())
}
\author{
Richèl J.C. Bilderbeek
}
