% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_scores_table.R
\name{check_scores_table}
\alias{check_scores_table}
\title{Check if the \code{scores_table} is valid}
\usage{
check_scores_table(scores_table)
}
\arguments{
\item{scores_table}{table that contains the dimensionality
reduction scores}
}
\value{
Nothing.
Will \link{stop} if \code{scores_table} is invalid.
}
\description{
Check if the \code{scores_table} is valid
Will \link{stop} if not.
}
\examples{
scores_table <- create_test_scores_table()
check_scores_table(scores_table)
}
\author{
Richèl J.C. Bilderbeek
}
