% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_train_times_table.R
\name{check_train_times_table}
\alias{check_train_times_table}
\title{Check if the \code{train_times_table} is valid}
\usage{
check_train_times_table(train_times_table)
}
\arguments{
\item{train_times_table}{a table with training time per epoch,
as can be checked with \link{check_train_times_table}}
}
\value{
Nothing.
Will \link{stop} if \code{train_times_table} is invalid.
}
\description{
Check if the \code{train_times_table} is valid
Will \link{stop} if not.
}
\examples{
train_times_table <- create_test_train_times_table()
check_train_times_table(train_times_table)
}
\author{
Richèl J.C. Bilderbeek
}
