% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_trainedmodelname.R
\name{check_trainedmodelname}
\alias{check_trainedmodelname}
\title{Check if the \code{trainedmodelname} is valid}
\usage{
check_trainedmodelname(trainedmodelname)
}
\arguments{
\item{trainedmodelname}{name of the trained model,
which is \code{ae_out} by default}
}
\value{
Nothing. Will \link{stop} if \code{trainedmodelname} is invalid.
}
\description{
Check if the \code{trainedmodelname} is valid
Will \link{stop} if not.
}
\examples{
check_trainedmodelname("ae_out")
}
\author{
Richèl J.C. Bilderbeek
}
