% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_file_path.R
\name{clean_file_path}
\alias{clean_file_path}
\title{Cleans up a file path}
\usage{
clean_file_path(file_path)
}
\arguments{
\item{file_path}{path to a file as a string}
}
\value{
a cleaned-up string
}
\description{
Cleans up a file path:
\itemize{
\item Remove double slashes
}
}
\examples{
# Does nothing on clean paths
clean_file_path("~/HO_superpopulations")

# Replace the double slashes by single slashes
clean_file_path("~//HO_superpopulations")
}
\author{
Richèl J.C. Bilderbeek
}
