% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_gcaer_tempfolder.R
\name{clean_gcaer_tempfolder}
\alias{clean_gcaer_tempfolder}
\title{Remove all \code{gcaer} temporary files}
\usage{
clean_gcaer_tempfolder()
}
\value{
Nothing
}
\description{
Remove all \code{gcaer} temporary files
}
\examples{
if (plinkr::is_on_ci()) {
  clean_gcaer_tempfolder()
}
}
\author{
Richèl J.C. Bilderbeek
}
