% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clone_gcae_repo.R
\name{clone_gcae_repo}
\alias{clone_gcae_repo}
\title{Clone the GCAE repo}
\usage{
clone_gcae_repo(
  gcae_options = create_gcae_options(),
  github_repo_url = get_default_github_repo_url(),
  github_repo_branch_name = get_default_github_repo_branch(),
  verbose = FALSE
)
}
\arguments{
\item{gcae_options}{options to run GCAE,
as created by \link{create_gcae_options}}

\item{github_repo_url}{URL of the GitHub repository,
e.g. \verb{https://github.com/kausmees/GenoCAE}
or \verb{https://github.com/cnettel/GenoCAE}
or \verb{https://github.com/richelbilderbeek/GenoCAE}.
Use \link{get_default_github_repo_url} for the default value}

\item{github_repo_branch_name}{name of the \code{git} branch,
e.g \code{master} or \code{Pheno}.
Use \link{get_default_github_repo_branch_name} for the default value}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link[plinkr]{check_verbose} to detect if this argument is valid.}
}
\value{
Nothing
}
\description{
Clone the GCAE repo
}
\author{
Richèl J.C. Bilderbeek
}
