% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_gcae_evaluate_args.R
\name{create_gcae_evaluate_args}
\alias{create_gcae_evaluate_args}
\title{Create the CLI arguments to evaluate the \code{GCAE} training}
\usage{
create_gcae_evaluate_args(gcae_setup, metrics, epoch)
}
\arguments{
\item{gcae_setup}{GCAE model setup,
as created by \link{create_gcae_setup}}
}
\value{
the CLI arguments
}
\description{
Create the CLI arguments to evaluate the \code{GCAE} training
}
\examples{
gcae_setup <- create_test_gcae_setup()
args <- create_gcae_evaluate_args(
  gcae_setup = gcae_setup
)
# 'python3' is the path to Python3
# 'run_gcae.py' is the path to the GenoCAE Python script
full_args <- c("python3", "run_gcae.py", args)
cat(full_args)
}
\author{
Richèl J.C. Bilderbeek
}
