% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_gcae_experiment_params.R
\name{create_gcae_experiment_params}
\alias{create_gcae_experiment_params}
\title{Create the parameters for a GCAE experiment}
\usage{
create_gcae_experiment_params(
  gcae_setup,
  analyse_epochs,
  metrics,
  gcae_options = create_gcae_options()
)
}
\arguments{
\item{gcae_setup}{GCAE model setup,
as created by \link{create_gcae_setup},
read from file by \link{read_gcae_setup_file},
checked by \link{check_gcae_setup},
and saved to file by \link{save_gcae_setup}.}

\item{analyse_epochs}{the epochs to analyse, e.g. \code{c(1, 2, 3)},
as checked by \link{check_analyse_epochs}}

\item{metrics}{the metrics, as checked by \link{check_metrics}.
Use \code{""} to indicate not to use a clusting metric}

\item{gcae_options}{options to run GCAE,
as created by \link{create_gcae_options}}
}
\value{
the parameters for a GCAE experiment
}
\description{
Create the parameters for a GCAE experiment
}
\examples{
create_gcae_experiment_params(
  gcae_setup = create_test_gcae_setup(),
  analyse_epochs = seq(1, 2, 3),
  metrics = "f1_score_3,f1_score_3"
)
}
\author{
Richèl J.C. Bilderbeek
}
