% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_gcae_input_filenames.R
\name{create_gcae_input_filenames}
\alias{create_gcae_input_filenames}
\title{Create a \code{gcae_input_filenames}}
\usage{
create_gcae_input_filenames(
  bed_filename,
  bim_filename,
  fam_filename,
  phe_filename,
  labels_filename
)
}
\arguments{
\item{labels_filename}{path to a \code{.csv} file containing super-
and -population labels, as can be read by \link{read_labels_table}}
}
\value{
a \code{gcae_input_filenames},
as can be checked by \link{check_gcae_input_filenames}
}
\description{
Create a \code{gcae_input_filenames},
as can be checked by \link{check_gcae_input_filenames}
}
\examples{
create_gcae_input_filenames(
  bed_filename = get_gcaer_filename("gcae_input_files_1.bed"),
  bim_filename = get_gcaer_filename("gcae_input_files_1.bim"),
  fam_filename = get_gcaer_filename("gcae_input_files_1.fam"),
  phe_filename = get_gcaer_filename("gcae_input_files_1.phe"),
  labels_filename = get_gcaer_filename("gcae_input_files_1_labels.csv")
)
}
\author{
Richèl J.C. Bilderbeek
}
