% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_gcae_input_files_1.R
\name{create_gcae_input_files_1}
\alias{create_gcae_input_files_1}
\title{Create setting 1 for GCAE}
\usage{
create_gcae_input_files_1(base_input_filename = "setting_1")
}
\arguments{
\item{base_input_filename}{base filename of the files to be created}
}
\value{
a \code{gcae_input_filenames}, as can be checked by
\link{check_gcae_input_filenames}
}
\description{
\itemize{
\item Date: 2022-01-17
\item Goal: do the simplest useful simulation possible
\item Hypothesis: on the dimensionality reduction plots,
there will be three points.
With 'points' I really mean points, i.e. not clusters,
as there are exactly 3 genotypes
\item n = 1000
}
}
\details{
\tabular{lllll}{
   Genotype \tab Phenotype \tab Allele frequency \tab Population \tab Superpopulation \cr
   AA \tab 0 \tab 25\% \tab A \tab Americas \cr
   AC \tab 50 \tab 50\% \tab B \tab Americas \cr
   CC \tab 100 \tab 25\% \tab C \tab Central/South Asia \cr
}
}
\examples{
# Create the files
gcae_input_filenames <- create_gcae_input_files_1()

# Clean up
file.remove(as.character(unlist(gcae_input_filenames)))
}
\author{
Richèl J.C. Bilderbeek
}
