% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_gcae_input_files_2.R
\name{create_gcae_input_files_2}
\alias{create_gcae_input_files_2}
\title{Create setting 2 for GCAE}
\usage{
create_gcae_input_files_2(
  base_input_filename = "setting_2",
  n_individuals = 1000,
  n_random_snps = 0,
  plink_options = plinkr::create_plink_options()
)
}
\arguments{
\item{base_input_filename}{base filename of the files to be created}

\item{n_individuals}{number of individuals}

\item{n_random_snps}{the number of SNPs that do nothing}

\item{plink_options}{PLINK options,
as created by \link[plinkr]{create_plink_options}}
}
\value{
a \code{gcae_input_filenames}, as can be checked by
\link{check_gcae_input_filenames}
}
\description{
\itemize{
\item Date: 2022-03-07
\item Goal: assess how well GCAE can predict a phenotype, where the genotype
is a simple relationship between the first SNP only and the phenotype
\item Hypothesis: it does not matter how many noisy SNPs there are,
prediction is always good. Performance decreases with more noise SNPs
\item n = 1000
}
}
\examples{
if (plinkr::is_plink_installed()) {
  # Create the files
  gcae_input_filenames <- create_gcae_input_files_2(
    base_input_filename = get_gcaer_tempfilename()
  )

  # Clean up
  file.remove(as.character(unlist(gcae_input_filenames)))
}
}
\author{
Richèl J.C. Bilderbeek
}
