% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_gcae_options.R
\name{create_gcae_options}
\alias{create_gcae_options}
\title{Create the options to run GCAE}
\usage{
create_gcae_options(gcae_version = "1.0")
}
\arguments{
\item{gcae_version}{version of GCAE, e.g. \code{"1.7"}
Use \link{get_gcae_version} to get the \code{GCAE} version.
Use \link{get_gcae_versions} to get all the supported \code{GCAE} versions.}
}
\value{
the GCAE options,
as can be checked by \link{check_gcae_options}
}
\description{
Create the options to run GCAE.
}
\examples{
# Use the default GCAE version
create_gcae_options()
}
\author{
Richèl J.C. Bilderbeek
}
