% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_gcae_setup.R
\name{create_gcae_setup}
\alias{create_gcae_setup}
\title{Create the GCAE setup}
\usage{
create_gcae_setup(
  datadir,
  data,
  superpops,
  model_id,
  train_opts_id,
  data_opts_id,
  pheno_model_id,
  trainedmodeldir = get_gcaer_tempfilename(pattern = "ae_out")
)
}
\arguments{
\item{datadir}{directory where sample data is stored.
If not absolute: assumed relative to \verb{GenoCAE/} directory.
A valid \code{datadir} ends with a slash (\code{/})
and does not start with a full stop (\code{.})}

\item{data}{file prefix, not including path, of the data files.
The data files must be in EIGENSTRAT
or PLINK binary (\code{.bed}, \code{.bim}, \code{.fam}) format)}

\item{superpops}{path to the \code{superpops} file,
which contains the labels table and can be read
using \link{read_labels_file}.
This argument is named after the GCAE \code{--superpops} CLI flag.}

\item{model_id}{model id, corresponding to a file models/model_id.json}

\item{train_opts_id}{train options id, corresponding to a file
train_opts/train_opts_id.json}

\item{data_opts_id}{data options id,
corresponding to a file data_opts/data_opts_id.json
The data options affect how data is input to the model.}

\item{pheno_model_id}{phenotype model ID,
as checked by \link{check_pheno_model_id}}

\item{trainedmodeldir}{name of the folder to save the trained model to.}
}
\value{
full paths to the created files
}
\description{
Create the GCAE setup
}
\examples{
create_gcae_setup(
  datadir = "my_datadir/",
  data = "HumanOrigins249_tiny",
  superpops = "HO_superpopulations",
  model_id = "M1",
  train_opts_id = "ex3",
  data_opts_id = "b_0_4",
  pheno_model_id = "p2",
  trainedmodeldir = "ae_out"
)
}
\author{
Richèl J.C. Bilderbeek
}
