% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_gcae_train_args.R
\name{create_gcae_train_args}
\alias{create_gcae_train_args}
\title{Create the CLI arguments to train with GCAE}
\usage{
create_gcae_train_args(gcae_setup, epochs, save_interval)
}
\arguments{
\item{gcae_setup}{GCAE model setup,
as created by \link{create_gcae_setup}}

\item{epochs}{number of epochs to train}

\item{save_interval}{epoch intervals at which to save state of model,
and at which to calculate the valid loss}
}
\value{
the CLI arguments
}
\description{
Create the CLI arguments to train with GCAE
}
\examples{
# Use the default GCAE version
create_gcae_options()
}
\author{
Richèl J.C. Bilderbeek
}
