% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_genotype_concordances_through_time_png.R
\name{create_genotype_concordances_through_time_png}
\alias{create_genotype_concordances_through_time_png}
\title{Create a PNG file of the genotype concordance through time}
\usage{
create_genotype_concordances_through_time_png(
  genotype_concordances_table,
  png_filename
)
}
\arguments{
\item{genotype_concordances_table}{genotype concordances
in time,
as can be checked by \link{check_genotype_concordances_table}}

\item{png_filename}{name of the \code{PNG} file to be produced,
as checked by \link{check_png_filename}}
}
\value{
a \link[ggplot2]{ggplot2}
}
\description{
Create a PNG file of the genotype concordance through time
}
\author{
Richèl J.C. Bilderbeek
}
