% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_losses_from_train_t_through_time_png.R
\name{create_losses_from_train_t_through_time_png}
\alias{create_losses_from_train_t_through_time_png}
\title{Create a PNG file of the losses from training through time,
based on the training set.}
\usage{
create_losses_from_train_t_through_time_png(
  losses_from_train_t_table,
  png_filename
)
}
\arguments{
\item{losses_from_train_t_table}{losses from training, measured
in the training step,
as can be checked by \link{check_losses_from_train_t_table}}

\item{png_filename}{name of the \code{PNG} file to be produced,
as checked by \link{check_png_filename}}
}
\value{
a \link[ggplot2]{ggplot2}
}
\description{
Create a PNG file of the losses from training through time,
based on the training set.
}
\author{
Richèl J.C. Bilderbeek
}
