% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_nmse_through_time_png.R
\name{create_nmse_through_time_png}
\alias{create_nmse_through_time_png}
\title{Create a PNG file of the NMSE through time}
\usage{
create_nmse_through_time_png(nmse_in_time_table, png_filename)
}
\arguments{
\item{nmse_in_time_table}{normalized mean squared error in time,
as can be checked by \link{check_nmse_in_time_table}}

\item{png_filename}{name of the \code{PNG} file to be produced,
as checked by \link{check_png_filename}}
}
\value{
a \link[ggplot2]{ggplot2}
}
\description{
Create a PNG file of the NMSE through time
}
\author{
Richèl J.C. Bilderbeek
}
