% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_phenotype_prediction_quality_png.R
\name{create_phenotype_prediction_quality_png}
\alias{create_phenotype_prediction_quality_png}
\title{Create a GIF file of the phenotype predictions through time}
\usage{
create_phenotype_prediction_quality_png(
  phenotype_predictions_table,
  png_filename
)
}
\arguments{
\item{phenotype_predictions_table}{a table of phenotype predictions,
with a true and estimated value for all individuals at a certain
epoch}

\item{png_filename}{name of the \code{PNG} file to be produced,
as checked by \link{check_png_filename}}
}
\value{
the result of \link[gganimate]{animate}
}
\description{
Create a GIF file of the phenotype predictions through time
}
\author{
Richèl J.C. Bilderbeek
}
