% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_phenotype_prediction_through_time_gif.R
\name{create_phenotype_prediction_through_time_gif}
\alias{create_phenotype_prediction_through_time_gif}
\title{Create a GIF file of the phenotype predictions through time}
\usage{
create_phenotype_prediction_through_time_gif(
  phenotype_predictions_table,
  gif_filename
)
}
\arguments{
\item{phenotype_predictions_table}{a table of phenotype predictions
with a true and estimated value for all individuals at a certain
epoch}

\item{gif_filename}{name of a GIF file,
as checked by \link{check_gif_filename}}
}
\value{
the result of \link[gganimate]{animate}
}
\description{
Create a GIF file of the phenotype predictions through time
}
\author{
Richèl J.C. Bilderbeek
}
