% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_phenotype_predictions_through_time_png.R
\name{create_phenotype_predictions_through_time_png}
\alias{create_phenotype_predictions_through_time_png}
\title{Create a PNG file of the phenotypic predictions through time}
\usage{
create_phenotype_predictions_through_time_png(
  phenotype_predictions_table,
  png_filename
)
}
\arguments{
\item{phenotype_predictions_table}{a table of phenotype predictions,
with a true and estimated value for all individuals at a certain
epoch}

\item{png_filename}{name of the \code{PNG} file to be produced,
as checked by \link{check_png_filename}}
}
\value{
a \link[ggplot2]{ggplot2}
}
\description{
Create a PNG file of the phenotypic predictions through time
}
\author{
Richèl J.C. Bilderbeek
}
