% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_plot_from_gcae_experiment_results.R
\name{create_plot_from_gcae_experiment_results}
\alias{create_plot_from_gcae_experiment_results}
\title{Create plots from the saved GCAE experiment's results}
\usage{
create_plot_from_gcae_experiment_results(folder_name)
}
\arguments{
\item{folder_name}{name of a folder,
as checked by \link{check_folder_name}}
}
\value{
filename of the saved plots
}
\description{
Create plots from the saved GCAE experiment's results
}
\examples{
create_test_gcae_setup()
check_gcae_setup(create_test_gcae_setup())
}
\author{
Richèl J.C. Bilderbeek
}
