% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_r_squared_through_time_png.R
\name{create_r_squared_through_time_png}
\alias{create_r_squared_through_time_png}
\title{Create a PNG file of the \code{r_squared} through time}
\usage{
create_r_squared_through_time_png(r_squared_in_time_table, png_filename)
}
\arguments{
\item{r_squared_in_time_table}{\code{r_squared} in time,
as can be checked by \link{check_r_squared_in_time_table}}

\item{png_filename}{name of the \code{PNG} file to be produced,
as checked by \link{check_png_filename}}
}
\value{
a \link[ggplot2]{ggplot2}
}
\description{
Create a PNG file of the \code{r_squared} through time
}
\author{
Richèl J.C. Bilderbeek
}
