% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_scores_through_time_png.R
\name{create_scores_through_time_png}
\alias{create_scores_through_time_png}
\title{Create a PNG file of the F1 scores through time}
\usage{
create_scores_through_time_png(scores_table, png_filename)
}
\arguments{
\item{png_filename}{name of the \code{PNG} file to be produced,
as checked by \link{check_png_filename}}
}
\value{
a \link[ggplot2]{ggplot2}
}
\description{
Create a PNG file of the F1 scores through time
}
\author{
Richèl J.C. Bilderbeek
}
