% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_test_gcae_experiment_results.R
\name{create_test_gcae_experiment_results}
\alias{create_test_gcae_experiment_results}
\title{Create a \code{gcae_experiment_results} to be used in testing}
\usage{
create_test_gcae_experiment_results()
}
\value{
a \code{gcae_experiment_results}
}
\description{
Create a \code{gcae_experiment_results} to be used in testing
}
\examples{
create_test_gcae_experiment_results()
check_gcae_experiment_results(create_test_gcae_experiment_results())
}
\author{
Richèl J.C. Bilderbeek
}
