% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_test_gcae_input_data.R
\name{create_test_gcae_input_data}
\alias{create_test_gcae_input_data}
\title{Create a \code{gcae_input_data} to be used in testing.}
\usage{
create_test_gcae_input_data()
}
\value{
a \code{gcae_input_data}
}
\description{
Create a \code{gcae_input_data} to be used in testing.
}
\examples{
gcae_input_data <- create_test_gcae_input_data()
check_gcae_input_data(gcae_input_data)
}
\author{
Richèl J.C. Bilderbeek
}
