% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_test_genotype_concordances_table.R
\name{create_test_genotype_concordances_table}
\alias{create_test_genotype_concordances_table}
\title{Create a \code{genotype_concordances_table} to be used in testing}
\usage{
create_test_genotype_concordances_table()
}
\value{
a \code{genotype_concordances_table}
}
\description{
Create a \code{genotype_concordances_table} to be used in testing
}
\examples{
create_test_genotype_concordances_table()
check_genotype_concordances_table(create_test_genotype_concordances_table())
}
\author{
Richèl J.C. Bilderbeek
}
