% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_test_losses_from_train_v_table.R
\name{create_test_losses_from_train_v_table}
\alias{create_test_losses_from_train_v_table}
\title{Create a \code{losses_from_train_v_table} to be used in testing}
\usage{
create_test_losses_from_train_v_table()
}
\value{
a \code{losses_from_train_v_table}
}
\description{
Create a \code{losses_from_train_v_table} to be used in testing
}
\examples{
create_test_losses_from_train_v_table()
check_losses_from_train_v_table(create_test_losses_from_train_v_table())
}
\author{
Richèl J.C. Bilderbeek
}
