% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_test_model.R
\name{create_test_model}
\alias{create_test_model}
\title{Create a GCAE model to be used in testing}
\usage{
create_test_model(model_filename = gcaer::get_gcaer_filename("M0.json"))
}
\arguments{
\item{model_filename}{name of a file that stores a \code{GCAE} architecture,
as checked by \link{check_model_filename}.
Use \link{read_model_file} to read a \code{GCAE} architecture from that file.}
}
\value{
a GCAE model
}
\description{
Create a GCAE model to be used in testing
}
\examples{
create_test_model()
check_model(create_test_model())
}
\author{
Richèl J.C. Bilderbeek
}
