% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_test_r_squared_in_time_table.R
\name{create_test_r_squared_in_time_table}
\alias{create_test_r_squared_in_time_table}
\title{Create a \code{r_squared_in_time_table} to be used in testing}
\usage{
create_test_r_squared_in_time_table()
}
\value{
a \code{r_squared_in_time_table}
}
\description{
Create a \code{r_squared_in_time_table} to be used in testing
}
\examples{
create_test_r_squared_in_time_table()
check_r_squared_in_time_table(create_test_r_squared_in_time_table())
}
\author{
Richèl J.C. Bilderbeek
}
