% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_test_score_per_pop_table.R
\name{create_test_score_per_pop_table}
\alias{create_test_score_per_pop_table}
\title{Create a \code{score_per_pop_table} to be used in testing}
\usage{
create_test_score_per_pop_table()
}
\value{
the parameters for a GCAE experiment
}
\description{
Create a \code{score_per_pop_table} to be used in testing
}
\examples{
check_score_per_pop_table(create_test_score_per_pop_table())
}
\author{
Richèl J.C. Bilderbeek
}
