% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_test_scores_table.R
\name{create_test_scores_table}
\alias{create_test_scores_table}
\title{Create a \code{scores_table} to be used in testing}
\usage{
create_test_scores_table()
}
\value{
a \code{scores_table}
}
\description{
Create a \code{scores_table} to be used in testing
}
\examples{
create_test_scores_table()
}
\author{
Richèl J.C. Bilderbeek
}
