% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_test_train_times_table.R
\name{create_test_train_times_table}
\alias{create_test_train_times_table}
\title{Create a \code{train_times_table} to be used in testing}
\usage{
create_test_train_times_table()
}
\value{
a \code{train_times_table}
}
\description{
Create a \code{train_times_table} to be used in testing
}
\examples{
create_test_train_times_table()
}
\author{
Richèl J.C. Bilderbeek
}
