% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_train_times_through_time_png.R
\name{create_train_times_through_time_png}
\alias{create_train_times_through_time_png}
\title{Create a PNG file of the training times through time}
\usage{
create_train_times_through_time_png(train_times_table, png_filename)
}
\arguments{
\item{train_times_table}{a table with training time per epoch,
as can be checked with \link{check_train_times_table}}

\item{png_filename}{name of the \code{PNG} file to be produced,
as checked by \link{check_png_filename}}
}
\value{
a \link[ggplot2]{ggplot2}
}
\description{
Create a PNG file of the training times through time
}
\author{
Richèl J.C. Bilderbeek
}
