% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_gcae_experiment.R
\name{do_gcae_experiment}
\alias{do_gcae_experiment}
\title{Do a full \code{GCAE} experiment}
\usage{
do_gcae_experiment(gcae_experiment_params, verbose = FALSE)
}
\arguments{
\item{gcae_experiment_params}{parameters to run a full
\code{GCAE} experimenr (i.e. call \link{do_gcae_experiment}),
as created by \link{create_gcae_experiment_params} and checked
by \link{check_gcae_experiment_params}}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link[plinkr]{check_verbose} to detect if this argument is valid.}
}
\value{
a \code{gcae_experiment_results}
}
\description{
Do a full \code{GCAE} experiment
}
\author{
Richèl J.C. Bilderbeek
}
