% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_phenotype_prediction.R
\name{evaluate_phenotype_prediction}
\alias{evaluate_phenotype_prediction}
\title{Determine how well the quantitative trait prediction worked}
\usage{
evaluate_phenotype_prediction(gcae_experiment_params, epoch, verbose = FALSE)
}
\arguments{
\item{gcae_experiment_params}{parameters to run a full
\code{GCAE} experimenr (i.e. call \link{do_gcae_experiment}),
as created by \link{create_gcae_experiment_params} and checked
by \link{check_gcae_experiment_params}}

\item{epoch}{the epoch, as checked by \link{check_epoch}}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link[plinkr]{check_verbose} to detect if this argument is valid.}
}
\value{
the plot
}
\description{
Determine how well the quantitative trait prediction worked
}
\author{
Richèl J.C. Bilderbeek
}
