% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcae_evaluate.R
\name{gcae_evaluate}
\alias{gcae_evaluate}
\title{Use \code{GCAE} its evaluate}
\usage{
gcae_evaluate(gcae_setup, gcae_options, metrics, epoch, verbose = FALSE)
}
\arguments{
\item{gcae_setup}{GCAE model setup,
as created by \link{create_gcae_setup},
read from file by \link{read_gcae_setup_file},
checked by \link{check_gcae_setup},
and saved to file by \link{save_gcae_setup}.}

\item{gcae_options}{options to run GCAE,
as created by \link{create_gcae_options}}

\item{metrics}{the metrics, as checked by \link{check_metrics}.
Use \code{""} to indicate not to use a clusting metric}

\item{epoch}{the epoch, as checked by \link{check_epoch}}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link[plinkr]{check_verbose} to detect if this argument is valid.}
}
\description{
Use \code{GCAE} its evaluate
}
\author{
Richèl J.C. Bilderbeek
}
