% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcae_options_to_str.R
\name{gcae_options_to_str}
\alias{gcae_options_to_str}
\title{Convert a \code{gcae_options} to a string}
\usage{
gcae_options_to_str(gcae_options)
}
\arguments{
\item{gcae_options}{options to run GCAE,
as created by \link{create_gcae_options}}
}
\value{
the \code{gcae_options} as a character vector,
with one string per \code{gcae_options} element
}
\description{
Convert a \code{gcae_options} to a string
}
\examples{
gcae_options_to_str(create_gcae_options())
}
\author{
Richèl J.C. Bilderbeek
}
