% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcae_project.R
\name{gcae_project}
\alias{gcae_project}
\title{Project}
\usage{
gcae_project(
  datadir = "example_tiny/",
  data = "HumanOrigins249_tiny",
  superpops = file.path(datadir, "HO_superpopulations"),
  gcae_setup = create_gcae_setup(),
  gcae_options = create_gcae_options(),
  verbose = FALSE
)
}
\arguments{
\item{datadir}{directory where sample data is stored}

\item{data}{file prefix, not including path, of the data files
(EIGENSTRAT of PLINK format)}

\item{superpops}{path to the superpops file.
This argument is named after the GCAE \code{--superpops} CLI flag.}

\item{gcae_setup}{GCAE model setup,
as created by \link{create_gcae_setup}}

\item{gcae_options}{options to run GCAE,
as created by \link{create_gcae_options}}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link{check_verbose} to detect if this argument is valid.}
}
\value{
full paths to the created files
}
\description{
Project, using the GCAE command-line interface
}
\author{
Richèl J.C. Bilderbeek
}
