% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcae_setup_to_str.R
\name{gcae_setup_to_str}
\alias{gcae_setup_to_str}
\title{Convert a \code{gcae_setup} to a string}
\usage{
gcae_setup_to_str(gcae_setup)
}
\arguments{
\item{gcae_setup}{GCAE model setup,
as created by \link{create_gcae_setup},
read from file by \link{read_gcae_setup_file},
checked by \link{check_gcae_setup},
and saved to file by \link{save_gcae_setup}.}
}
\value{
the \code{gcae_setup} as a character vector,
with one string per \code{gcae_setup} element
}
\description{
Convert a \code{gcae_setup} to a string
}
\examples{
gcae_setup_to_str(create_test_gcae_setup())
}
\author{
Richèl J.C. Bilderbeek
}
