% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcae_train_more.R
\name{gcae_train_more}
\alias{gcae_train_more}
\title{Continue training a model, and save its state at certain epochs to disk.}
\usage{
gcae_train_more(
  gcae_setup = create_test_gcae_setup(),
  resume_from = 0,
  epochs = 1,
  save_interval = 1,
  gcae_options = create_gcae_options(),
  verbose = FALSE
)
}
\arguments{
\item{gcae_setup}{GCAE model setup,
as created by \link{create_gcae_setup},
read from file by \link{read_gcae_setup_file},
checked by \link{check_gcae_setup},
and saved to file by \link{save_gcae_setup}.}

\item{resume_from}{the epoch to resume from,
as checked by \link{chgeck_resume_from}}

\item{epochs}{number of epochs, as checked by \link{check_epochs}}

\item{save_interval}{epoch intervals at which to save state of model,
and at which to calculate the valid loss}

\item{gcae_options}{options to run GCAE,
as created by \link{create_gcae_options}}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link[plinkr]{check_verbose} to detect if this argument is valid.}
}
\value{
full paths to the created files
}
\description{
Continue training a model, and save its state at certain epochs to disk.
use the GCAE command-line interface.
}
\details{
To simply run the model from epoch zero,
use \link{gcae_train}
}
\seealso{
Use \link{gcae_train} to do train from epoch zero to end in one go.
Use \link{create_gcae_train_more_args} to get the command-line
arguments.
}
\author{
Richèl J.C. Bilderbeek
}
