% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcaer.R
\docType{package}
\name{gcaer}
\alias{gcaer}
\title{\link{gcaer}: An R interface to \code{GCAE}.}
\description{
\link{gcaer} is an R interface to \code{GCAE}.
}
\details{
\link{gcaer} allows the user to install, run and uninstall
(multiple versions of) \code{GCAE}.
Also GCAE output files can be read, modified within R
and saved to GCAE-conforming files

This is a list of functions to install or uninstall \code{GCAE}:
\itemize{
\item \link{check_gcae_is_installed}:
check if \code{GCAE} is installed, stop if not
\item \link{is_gcae_installed}:
determine if \code{GCAE} is installed, yes/no
\item \link{get_gcae_version}: get the version of \code{GCAE}
\item \link{install_gcae}: install \code{GCAE}
\item \link{uninstall_gcae}: uninstall \code{GCAE}
}

This is a list of functions to run or with with
(multiple versions of) \code{GCAE}:
\itemize{
\item \link{run_gcae}: run \code{GCAE}
}
}
\examples{
is_gcae_installed()
}
\author{
Richèl J.C. Bilderbeek
}
