% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcaer_report.R
\name{gcaer_report}
\alias{gcaer_report}
\title{Create a \link{gcaer} report, to be used when reporting bugs}
\usage{
gcaer_report(gcae_options = gcaer::create_gcae_options(), verbose = FALSE)
}
\arguments{
\item{gcae_options}{options to run GCAE,
as created by \link{create_gcae_options}}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link[plinkr]{check_verbose} to detect if this argument is valid.}
}
\description{
Create a \link{gcaer} report, to be used when reporting bugs
}
\examples{
if (plinkr::is_on_ci()) {
  gcaer_report()
}
}
\author{
Richèl J.C. Bilderbeek
}
