% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_conda_binary_path.R
\name{get_conda_binary_path}
\alias{get_conda_binary_path}
\title{Get the path to the \code{conda} binary}
\usage{
get_conda_binary_path(gcae_options = create_gcae_options())
}
\arguments{
\item{gcae_options}{options to run GCAE,
as created by \link{create_gcae_options}}
}
\value{
the path to the \code{conda} binary
}
\description{
Get the path to the \code{conda} binary
}
\examples{
get_conda_binary_path()
}
\author{
Richèl J.C. Bilderbeek
}
