% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_default_github_repo_branch_name.R
\name{get_default_github_repo_branch_name}
\alias{get_default_github_repo_branch_name}
\title{Get the name of the default GitHub repository branch}
\usage{
get_default_github_repo_branch_name()
}
\value{
a string with the name of the git branch
}
\description{
Get the name of the default GitHub repository branch
}
\examples{
get_default_github_repo_branch()
}
\author{
Richèl J.C. Bilderbeek
}
