% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_default_github_repo_url.R
\name{get_default_github_repo_url}
\alias{get_default_github_repo_url}
\title{Get the default GitHub repository URL}
\usage{
get_default_github_repo_url()
}
\value{
a string with the URL
}
\description{
Get the default GitHub repository URL
}
\examples{
get_default_github_repo_url()
}
\author{
Richèl J.C. Bilderbeek
}
