% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_default_ormr_folder_name.R
\name{get_default_ormr_folder_name}
\alias{get_default_ormr_folder_name}
\title{Get the default \link[ormr]{ormr} folder used by \code{gcaer}}
\usage{
get_default_ormr_folder_name()
}
\value{
the default \link[ormr]{ormr} folder used by \code{gcaer}
}
\description{
Get the default \link[ormr]{ormr} folder used by \code{gcaer}
}
\examples{
get_default_ormr_folder_name()
}
\author{
Richèl J.C. Bilderbeek
}
