% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gcae_data_opts_filename.R
\name{get_gcae_data_opts_filename}
\alias{get_gcae_data_opts_filename}
\title{Get the filename to the GCAE data options}
\usage{
get_gcae_data_opts_filename(data_opts_id, gcae_options = create_gcae_options())
}
\arguments{
\item{data_opts_id}{data options id,
corresponding to a file data_opts/data_opts_id.json
The data options affect how data is input to the model.}

\item{gcae_options}{options to run GCAE,
as created by \link{create_gcae_options}}
}
\value{
full path to the data options' JSON file
}
\description{
Get the filename to the GCAE data options.
This function is named after the GCAE \code{data_opts} folder.
}
\examples{
get_gcae_data_opts_filename(
  data_opts_id = "b_0_4"
)
}
\seealso{
Use \link{get_gcae_data_opts_filenames} to get all the
files for the data storage options.

Use \link{get_gcae_data_opts_ids} to get the IDs
of all the data storage options.

To obtain the paths to the files for the different \code{GCAE} features:
\itemize{
\item Autoencoder architecture: use \link{get_gcae_model_filename}
\item Phenotypic prediction: use \link{get_gcae_pheno_model_filename}
\item Training options: use \link{get_gcae_train_opts_filename}
\item File storage options: use \link{get_gcae_data_opts_filename}
}
}
\author{
Richèl J.C. Bilderbeek
}
