% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gcae_data_opts_filenames.R
\name{get_gcae_data_opts_filenames}
\alias{get_gcae_data_opts_filenames}
\title{Get the filenames for all the \code{GCAE} phenotypic model architectures}
\usage{
get_gcae_data_opts_filenames(gcae_options = create_gcae_options())
}
\arguments{
\item{gcae_options}{options to run GCAE,
as created by \link{create_gcae_options}}
}
\value{
full paths to all the \code{GCAE} phenotypic model architectures filenames
}
\description{
Get the filenames for all the \code{GCAE} phenotypic model architectures.
}
\examples{
get_gcae_data_opts_filenames()
}
\seealso{
use \link{get_gcae_model_filename} to get the filename
of one \code{GCAE} autoencoder architecture

Use \link{get_gcae_model_filenames} to get the filenames for the
autoencoder models.
}
\author{
Richèl J.C. Bilderbeek
}
