% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gcae_example_filename.R
\name{get_gcae_example_filename}
\alias{get_gcae_example_filename}
\title{Get the path to a gcaer file}
\usage{
get_gcae_example_filename(
  example_filename,
  gcae_options = create_gcae_options()
)
}
\arguments{
\item{example_filename}{name of the example file}

\item{gcae_options}{options to run GCAE,
as created by \link{create_gcae_options}}
}
\value{
the path to a gcaer file, if it exists.
Will \link{stop} if the file does not exist
}
\description{
Get the path to a gcaer file
}
\examples{
if (plinkr::is_on_ci()) {

  if (has_cloned_gcae_repo()) { # is_gcae_installed is too slow
    get_gcae_example_filename("HumanOrigins249_tiny.fam")
  }
}
}
\author{
Richèl J.C. Bilderbeek
}
