% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gcae_example_filenames.R
\name{get_gcae_example_filenames}
\alias{get_gcae_example_filenames}
\title{Get the paths to all \code{GCAE} example files}
\usage{
get_gcae_example_filenames(gcae_options = create_gcae_options())
}
\arguments{
\item{gcae_options}{options to run GCAE,
as created by \link{create_gcae_options}}
}
\value{
the path to all \code{GCAE} files
}
\description{
Get the paths to all \code{GCAE} example files
}
\examples{
if (is_gcae_installed()) {
  get_gcae_example_filenames()
}
}
\author{
Richèl J.C. Bilderbeek
}
