% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gcae_experiment_results_filenames.R
\name{get_gcae_experiment_results_filenames}
\alias{get_gcae_experiment_results_filenames}
\title{Get a list of filenames the GCAE results will be saved to}
\usage{
get_gcae_experiment_results_filenames(folder_name, file_extension = ".csv")
}
\arguments{
\item{folder_name}{name of a folder,
as checked by \link{check_folder_name}}

\item{file_extension}{a file extension, for example \code{.R}
or \code{.csv}.}
}
\value{
Nothing
}
\description{
Get a list of filenames the GCAE results will be saved to
}
\examples{
get_gcae_experiment_results_filenames(folder_name = tempdir())
}
\author{
Richèl J.C. Bilderbeek
}
