% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gcae_folder.R
\name{get_gcae_folder}
\alias{get_gcae_folder}
\title{Get the path to the default \code{GCAE} folder as used by \link{gcaer}}
\usage{
get_gcae_folder()
}
\value{
path to the default \code{GCAE} folder as used by \link{gcaer}
}
\description{
Get the path to the default \code{GCAE} folder as used by \link{gcaer}.
This is the folder where all \code{GCAE} versions are installed.
}
\details{
Use \link{get_gcaer_folder} to get the folder where \link{gcaer}
stores its temporary files
}
\author{
Richèl J.C. Bilderbeek
}
