% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gcae_input_filenames.R
\name{get_gcae_input_filenames}
\alias{get_gcae_input_filenames}
\title{Create a \code{gcae_input_filenames}}
\usage{
get_gcae_input_filenames(gcae_setup)
}
\arguments{
\item{gcae_setup}{GCAE model setup,
as created by \link{create_gcae_setup},
read from file by \link{read_gcae_setup_file},
checked by \link{check_gcae_setup},
and saved to file by \link{save_gcae_setup}.}
}
\value{
a \code{gcae_input_filenames},
as can be checked by \link{check_gcae_input_filenames}
}
\description{
Create a \code{gcae_input_filenames},
as can be checked by \link{check_gcae_input_filenames},
from a \code{gcae_setup} (see \link{check_gcae_setup})
and a \code{gcae_options} (see \link{check_gcae_options})
using default naming
}
\examples{
gcae_setup <- create_test_gcae_setup()
gcae_input_filenames <- get_gcae_input_filenames(
  gcae_setup = gcae_setup
)
file.exists(gcae_input_filenames$bed_filename)
file.exists(gcae_input_filenames$bim_filename)
file.exists(gcae_input_filenames$fam_filename)
file.exists(gcae_input_filenames$phe_filename)
file.exists(gcae_input_filenames$labels_filename)
}
\seealso{
use \link{create_gcae_input_filenames} to
create the filesnames yourself
}
\author{
Richèl J.C. Bilderbeek
}
