% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gcae_model_filename.R
\name{get_gcae_model_filename}
\alias{get_gcae_model_filename}
\title{Get the filename to the GCAE model}
\usage{
get_gcae_model_filename(model_id = "M0", gcae_options = create_gcae_options())
}
\arguments{
\item{model_id}{model id, corresponding to a file models/model_id.json}

\item{gcae_options}{options to run GCAE,
as created by \link{create_gcae_options}}
}
\value{
full paths to the created files
}
\description{
Get the filename to the GCAE model.
This function is named after the GCAE \code{models} folder.
}
\examples{
get_gcae_model_filename()
get_gcae_model_filename(model_id = "M0")
}
\seealso{
use \link{get_gcae_pheno_model_filename} to get the filename
to a phenotype model.
}
\author{
Richèl J.C. Bilderbeek
}
