% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gcae_model_filenames.R
\name{get_gcae_model_filenames}
\alias{get_gcae_model_filenames}
\title{Get the filenames for all the \code{GCAE} autoencoder architectures}
\usage{
get_gcae_model_filenames(gcae_options = create_gcae_options())
}
\arguments{
\item{gcae_options}{options to run GCAE,
as created by \link{create_gcae_options}}
}
\value{
full paths to all the \code{GCAE} autoencoder architectures filenames
}
\description{
Get the filenames for all the \code{GCAE} autoencoder architectures.
}
\details{
This function is named after the GCAE \code{models} folder,
although the phenotypic prediction models
(that can be obtained using \link{get_gcae_pheno_model_filenames})
are stored there as well.
}
\examples{
get_gcae_model_filenames()
}
\seealso{
use \link{get_gcae_model_filename} to get the filename
of one \code{GCAE} autoencoder architecture

Use \link{get_gcae_pheno_model_filenames} to get the filenames for the
phenotypic models.
}
\author{
Richèl J.C. Bilderbeek
}
