% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gcae_model_ids.R
\name{get_gcae_model_ids}
\alias{get_gcae_model_ids}
\title{Get the model IDs for all the \code{GCAE} autoencoder architectures}
\usage{
get_gcae_model_ids(gcae_options = create_gcae_options())
}
\arguments{
\item{gcae_options}{options to run GCAE,
as created by \link{create_gcae_options}}
}
\value{
the model IDs of the \code{GCAE} autoencoder architectures
}
\description{
Get the model IDs for all the \code{GCAE} autoencoder architectures.
}
\details{
This function is named after the GCAE \code{models} folder,
although the phenotypic prediction models
(that can be obtained using \link{get_gcae_pheno_model_ids})
are stored there as well.
}
\examples{
get_gcae_model_ids()
}
\seealso{
use \link{get_gcae_model_id} to get the filename
of one \code{GCAE} autoencoder architecture

Use \link{get_gcae_pheno_model_ids} to get the IDs for the
phenotypic models.
}
\author{
Richèl J.C. Bilderbeek
}
