% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gcae_output_folder.R
\name{get_gcae_output_folder}
\alias{get_gcae_output_folder}
\title{Get the path to the GCAE output folder}
\usage{
get_gcae_output_folder(gcae_options = create_gcae_options())
}
\arguments{
\item{gcae_options}{options to run GCAE,
as created by \link{create_gcae_options}}
}
\value{
full paths to the created files
}
\description{
Get the path to the GCAE output folder.
This folder contains the output of multiple models.
}
\examples{
get_gcae_output_folder()
get_gcae_output_folder(create_gcae_options())
}
\author{
Richèl J.C. Bilderbeek
}
