% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gcae_output_subfolder.R
\name{get_gcae_output_subfolder}
\alias{get_gcae_output_subfolder}
\title{Get the path to the GCAE output subfolder}
\usage{
get_gcae_output_subfolder(gcae_setup)
}
\arguments{
\item{gcae_setup}{GCAE model setup,
as created by \link{create_gcae_setup}}
}
\value{
full paths to the created files
}
\description{
Get the path to the GCAE output subfolder.
This subfolder contains the output of one model
}
\examples{
get_gcae_output_subfolder(
  gcae_setup = create_gcae_setup(
    datadir = "my_datadir/",
    data = "HumanOrigins249_tiny"
  )
)
}
\author{
Richèl J.C. Bilderbeek
}
