% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gcae_pheno_model_filename.R
\name{get_gcae_pheno_model_filename}
\alias{get_gcae_pheno_model_filename}
\title{Get the filename to the \code{GCAE} phenotype model}
\usage{
get_gcae_pheno_model_filename(
  pheno_model_id = "p0",
  gcae_options = create_gcae_options()
)
}
\arguments{
\item{pheno_model_id}{phenotype model ID,
as checked by \link{check_pheno_model_id}}

\item{gcae_options}{options to run GCAE,
as created by \link{create_gcae_options}}
}
\value{
full paths to the created files
}
\description{
Get the filename to the \code{GCAE} model.
This function is named after the GCAE \code{pheno_model_id} flag.
}
\examples{
get_gcae_pheno_model_filename()
get_gcae_pheno_model_filename(pheno_model_id = "p0")
get_gcae_pheno_model_filename(pheno_model_id = "p1")
get_gcae_pheno_model_filename(pheno_model_id = "p2")
}
\seealso{
use \link{get_gcae_model_filename} to get the filename
to a model.
}
\author{
Richèl J.C. Bilderbeek
}
