% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gcae_required_python_packages.R
\name{get_gcae_required_python_packages}
\alias{get_gcae_required_python_packages}
\title{Get the names and version fo all the Python packages GCAE needs}
\usage{
get_gcae_required_python_packages(gcae_options = create_gcae_options())
}
\arguments{
\item{gcae_options}{options to run GCAE,
as created by \link{create_gcae_options}}
}
\value{
a character vector
}
\description{
Get the names and version fo all the Python packages GCAE needs
}
