% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gcae_subfolder.R
\name{get_gcae_subfolder}
\alias{get_gcae_subfolder}
\title{Get the GCAE subfolder}
\usage{
get_gcae_subfolder(gcae_options = create_gcae_options())
}
\arguments{
\item{gcae_options}{options to run GCAE,
as created by \link{create_gcae_options}}
}
\description{
Get the GCAE subfolder
}
