% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gcae_train_opts_filename.R
\name{get_gcae_train_opts_filename}
\alias{get_gcae_train_opts_filename}
\title{Get the filename to the GCAE training options}
\usage{
get_gcae_train_opts_filename(
  train_opts_id,
  gcae_options = create_gcae_options()
)
}
\arguments{
\item{train_opts_id}{train options id, corresponding to a file
train_opts/train_opts_id.json}

\item{gcae_options}{options to run GCAE,
as created by \link{create_gcae_options}}
}
\value{
full paths to the created files
}
\description{
Get the filename to the GCAE training options.
This function is named after the GCAE \code{train_opts} folder.
}
\seealso{
Use \link{get_gcae_train_opts_filenames} to get all the
files with training options.

To obtain the paths to the files for the different \code{GCAE} features:
\itemize{
\item Autoencoder architecture: use \link{get_gcae_model_filename}
\item Phenotypic preditcion: use \link{get_gcae_pheno_model_filename}
\item Training options: use \link{get_gcae_train_opts_filename}
\item File storage options: use \link{get_gcae_data_opts_filename}
}
}
\author{
Richèl J.C. Bilderbeek
}
