% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gcae_train_opts_ids.R
\name{get_gcae_train_opts_ids}
\alias{get_gcae_train_opts_ids}
\title{Get the \code{train_opts} IDs for all the \code{GCAE} phenotype architectures}
\usage{
get_gcae_train_opts_ids(gcae_options = create_gcae_options())
}
\arguments{
\item{gcae_options}{options to run GCAE,
as created by \link{create_gcae_options}}
}
\value{
the \code{train_opts} IDs of the \code{GCAE} phenotype architectures
}
\description{
Get the \code{train_opts} IDs for all the \code{GCAE} phenotype architectures.
}
\examples{
get_gcae_train_opts_ids()
}
\seealso{
Use \link{get_gcae_model_ids} to get the IDs for the
autoencoder models.
}
\author{
Richèl J.C. Bilderbeek
}
