% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gcae_trainedmodelname.R
\name{get_gcae_trainedmodelname}
\alias{get_gcae_trainedmodelname}
\title{Get the GCAE \code{trainedmodelname}}
\usage{
get_gcae_trainedmodelname(gcae_setup)
}
\arguments{
\item{gcae_setup}{GCAE model setup,
as created by \link{create_gcae_setup},
read from file by \link{read_gcae_setup_file},
checked by \link{check_gcae_setup},
and saved to file by \link{save_gcae_setup}.}
}
\value{
the GCAE trained model name
}
\description{
Get the GCAE trained model name,
to be used by the GCAE\code{--trainedmodelname} CLI flag
}
\examples{
gcae_setup <- create_test_gcae_setup()
get_gcae_trainedmodelname(gcae_setup)
}
\author{
Richèl J.C. Bilderbeek
}
