% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gcae_version.R
\name{get_gcae_version}
\alias{get_gcae_version}
\title{Get the version of GCAE}
\usage{
get_gcae_version(gcae_options = create_gcae_options())
}
\arguments{
\item{gcae_options}{options to run GCAE,
as created by \link{create_gcae_options}}
}
\value{
the version of GCAE
}
\description{
Get the version of GCAE
}
\author{
Richèl J.C. Bilderbeek
}
