% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gcaer_filename.R
\name{get_gcaer_filename}
\alias{get_gcaer_filename}
\title{Get the path to a gcaer file}
\usage{
get_gcaer_filename(example_filename)
}
\arguments{
\item{example_filename}{name of the example file}
}
\value{
the path to a gcaer file, if it exists.
Will \link{stop} if the file does not exist
}
\description{
Get the path to a gcaer file
}
\examples{
if (1 == 2) get_gcaer_filename("gcae.log")
}
\author{
Richèl J.C. Bilderbeek
}
