% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gcaer_filenames.R
\name{get_gcaer_filenames}
\alias{get_gcaer_filenames}
\title{Get the full paths to all gcaer files}
\usage{
get_gcaer_filenames()
}
\value{
the full paths to all gcaer files
}
\description{
Get the full paths to all gcaer files
}
\examples{
get_gcaer_filenames()
}
\author{
Richèl J.C. Bilderbeek
}
