% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gcaer_folder.R
\name{get_gcaer_folder}
\alias{get_gcaer_folder}
\title{Get the folder where \link{gcaer} stores its temporary files}
\usage{
get_gcaer_folder()
}
\value{
path to the default \code{gcaer} cache folder
}
\description{
Get the folder where \link{gcaer} stores its temporary files.
}
\details{
Use \code{get_gcae_folder} to get the path to the folder when GCAE is installed.
}
\author{
Richèl J.C. Bilderbeek
}
