% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gcaer_theme.R
\name{get_gcaer_theme}
\alias{get_gcaer_theme}
\title{Get the \code{gcaer} theme}
\usage{
get_gcaer_theme()
}
\value{
a \link[ggplot2]{theme}
}
\description{
Get the \code{gcaer} theme
}
\author{
Richèl J.C. Bilderbeek
}
