% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_miniconda_path.R
\name{get_miniconda_path}
\alias{get_miniconda_path}
\title{Get the path where Miniconda and all Pyton packages will be installed}
\usage{
get_miniconda_path(gcae_options = create_gcae_options())
}
\arguments{
\item{gcae_options}{options to run GCAE,
as created by \link{create_gcae_options}}
}
\value{
the \code{miniconda_path}, as used by \link[reticulate]{reticulate}
}
\description{
Get the path where Miniconda and all Pyton packages will be installed
}
\examples{
get_miniconda_path()
}
\author{
Richèl J.C. Bilderbeek
}
