% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_n_neurons_in_latent_layer.R
\name{get_n_neurons_in_latent_layer}
\alias{get_n_neurons_in_latent_layer}
\title{Get the number of neurons in the layent layer}
\usage{
get_n_neurons_in_latent_layer(model)
}
\value{
the number of neurons in the latent layer
}
\description{
Get the number of neurons in the layent layer
}
\examples{
model_filename <- get_gcae_model_filename("M1")
model <- read_model_file(model_filename)
n_neurons <- get_n_neurons_in_latent_layer(model)
}
\author{
Richèl J.C. Bilderbeek
}
