% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/get_n_neurons_in_latent_layer_from_gcae_experiment_params.R
\name{get_n_neurons_in_latent_layer_from_gcae_experiment_params}
\alias{get_n_neurons_in_latent_layer_from_gcae_experiment_params}
\title{Get the number of neurons in the layent layer from a \code{gcae_experiment_params}}
\usage{
get_n_neurons_in_latent_layer_from_gcae_experiment_params(
  gcae_experiment_params
)
}
\arguments{
\item{gcae_experiment_params}{parameters to run a full
\code{GCAE} experimenr (i.e. call \link{do_gcae_experiment}),
as created by \link{create_gcae_experiment_params} and checked
by \link{check_gcae_experiment_params}}
}
\value{
the number of neurons in the latent layer
}
\description{
Get the number of neurons in the layent layer from a \code{gcae_experiment_params}
}
\examples{
get_n_neurons_in_latent_layer_from_gcae_experiment_params(
  create_test_gcae_experiment_params()
)
}
\seealso{
use \link{get_n_neurons_in_latent_layer}
to get the number of neurons in the latent layer
for different input arguments
}
\author{
Richèl J.C. Bilderbeek
}
