% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_run_gcae_py_path.R
\name{get_run_gcae_py_path}
\alias{get_run_gcae_py_path}
\title{Get the full path to \code{run_gcae.py}}
\usage{
get_run_gcae_py_path(gcae_options)
}
\arguments{
\item{gcae_options}{options to run GCAE,
as created by \link{create_gcae_options}}
}
\value{
the full path to \code{run_gcae.py}
}
\description{
Get the full path to \code{run_gcae.py}
}
\author{
Richèl J.C. Bilderbeek
}
