% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_test_data.R
\name{get_test_data}
\alias{get_test_data}
\title{Get the \code{data} for use in testing.}
\usage{
get_test_data()
}
\value{
a \code{data}
}
\description{
Get the \code{data} (as can be checked by \link{check_data})
for use in testing.
}
\examples{
get_test_data()
}
\author{
Richèl J.C. Bilderbeek
}
