% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_test_datadir.R
\name{get_test_datadir}
\alias{get_test_datadir}
\title{Get the \code{datadir} for use in testing.}
\usage{
get_test_datadir()
}
\value{
a \code{datadir}
}
\description{
Get the \code{datadir} (as can be checked by \link{check_datadir})
for use in testing.
}
\examples{
get_test_datadir()
}
\author{
Richèl J.C. Bilderbeek
}
