% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_test_labels_table.R
\name{get_test_labels_table}
\alias{get_test_labels_table}
\title{Get a \code{labels_table} for use in testing.}
\usage{
get_test_labels_table()
}
\value{
a \code{labels_table}
}
\description{
Get a \code{labels_table} (as can be checked by \link{check_labels_table})
for use in testing.
}
\examples{
get_test_labels_table()
}
\author{
Richèl J.C. Bilderbeek
}
