% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_cloned_gcae_repo.R
\name{has_cloned_gcae_repo}
\alias{has_cloned_gcae_repo}
\title{Determines if the GCAE repo has been cloned already}
\usage{
has_cloned_gcae_repo(gcae_options = create_gcae_options())
}
\arguments{
\item{gcae_options}{options to run GCAE,
as created by \link{create_gcae_options}}
}
\value{
TRUE if the GCAE repo has been cloned already
}
\description{
Determines if the GCAE repo has been cloned already
}
\examples{
has_cloned_gcae_repo()
}
\seealso{
use \link{clone_gcae_repo} to clone the GCAE repo
}
\author{
Richèl J.C. Bilderbeek
}
