% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_gcae.R
\name{install_gcae}
\alias{install_gcae}
\title{Install a specific version of GCAE}
\usage{
install_gcae(gcae_options = create_gcae_options())
}
\arguments{
\item{gcae_options}{options to run GCAE,
as created by \link{create_gcae_options}}
}
\value{
Nothing
}
\description{
Install a specific version of GCAE
}
\author{
Richèl J.C. Bilderbeek
}
