% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_gcae_requirements.R
\name{install_gcae_requirements}
\alias{install_gcae_requirements}
\title{Install the GCAE requirements}
\usage{
install_gcae_requirements(
  gcae_options = create_gcae_options(),
  verbose = FALSE
)
}
\arguments{
\item{gcae_options}{options to run GCAE,
as created by \link{create_gcae_options}}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link[plinkr]{check_verbose} to detect if this argument is valid.}
}
\description{
Install the GCAE requirements
}
