% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_equal_json.R
\name{is_equal_json}
\alias{is_equal_json}
\title{Determine if two JSON objects are the same}
\usage{
is_equal_json(json_1, json_2)
}
\arguments{
\item{json_1}{a JSON object, as returned by \link[jsonlite]{read_json}}

\item{json_2}{a JSON object, as returned by \link[jsonlite]{read_json}}
}
\value{
\link{TRUE} if the objects are the same
}
\description{
Determine if two JSON objects are the same
}
\examples{
# Using an example file
example_model <- read_model_file(get_gcaer_filename("M0.json"))
is_equal_json(example_model, example_model)

# Using GCAE files
if (is_gcae_repo_cloned()) {
  m0 <- read_model_file(get_gcae_model_filename("M0"))
  m1 <- read_model_file(get_gcae_model_filename("M1"))
  is_equal_json(m0, m1)
}
}
\author{
Richèl J.C. Bilderbeek
}
