% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_gcae_experiment_params.R
\name{is_gcae_experiment_params}
\alias{is_gcae_experiment_params}
\title{Determine if the \code{gcae_experiment_params} is indeed
a \code{gcae_experiment_params}}
\usage{
is_gcae_experiment_params(gcae_experiment_params, verbose = FALSE)
}
\arguments{
\item{gcae_experiment_params}{parameters to run a full
\code{GCAE} experimenr (i.e. call \link{do_gcae_experiment}),
as created by \link{create_gcae_experiment_params} and checked
by \link{check_gcae_experiment_params}}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link[plinkr]{check_verbose} to detect if this argument is valid.}
}
\value{
\link{TRUE} if the \code{gcae_experiment_params} is indeed
a \code{gcae_experiment_params}
}
\description{
Determine if the \code{gcae_experiment_params} is indeed
a \code{gcae_experiment_params}
}
\examples{
# TRUE
is_gcae_experiment_params(create_test_gcae_experiment_params())

# FALSE
is_gcae_experiment_params(gcae_experiment_params = "nonsense")
is_gcae_experiment_params("nonsense", verbose = TRUE)
}
\author{
Richèl J.C. Bilderbeek
}
