% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_gcae_installed.R
\name{is_gcae_installed}
\alias{is_gcae_installed}
\title{Determine if \code{GCAE} is installed}
\usage{
is_gcae_installed()
}
\value{
\code{TRUE} if \code{GCAE} is installed
}
\description{
Determine if \code{GCAE} is installed
}
\examples{
is_gcae_installed()
}
\author{
Richèl J.C. Bilderbeek
}
