% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_gcae_repo_cloned.R
\name{is_gcae_repo_cloned}
\alias{is_gcae_repo_cloned}
\title{Determine if \code{GCAE}'s repo is cloned}
\usage{
is_gcae_repo_cloned(gcae_options = create_gcae_options(), verbose = FALSE)
}
\arguments{
\item{gcae_options}{options to run GCAE,
as created by \link{create_gcae_options}}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link[plinkr]{check_verbose} to detect if this argument is valid.}
}
\value{
\code{TRUE} if \code{GCAE}'s repo is cloned
}
\description{
Determine if \code{GCAE}'s repo is cloned
}
\examples{
is_gcae_repo_cloned()
}
\author{
Richèl J.C. Bilderbeek
}
