% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_gcae_script_fixed.R
\name{is_gcae_script_fixed}
\alias{is_gcae_script_fixed}
\title{Clone the GCAE repo}
\usage{
is_gcae_script_fixed(gcae_options = create_gcae_options(), verbose = FALSE)
}
\arguments{
\item{gcae_options}{options to run GCAE,
as created by \link{create_gcae_options}}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link[plinkr]{check_verbose} to detect if this argument is valid.}
}
\value{
Nothing
}
\description{
Clone the GCAE repo
}
\author{
Richèl J.C. Bilderbeek
}
