% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_miniconda_installed.R
\name{is_miniconda_installed}
\alias{is_miniconda_installed}
\title{Determine if Miniconda is installed}
\usage{
is_miniconda_installed(
  miniconda_path = get_miniconda_path(create_gcae_options())
)
}
\arguments{
\item{miniconda_path}{the Miniconda path,
as can be obtained using \link{get_miniconda_path}
of a \code{gcae_options} (see \link{create_gcae_options})}
}
\value{
TRUE if Miniconda is installed
}
\description{
Determine if Miniconda is installed
}
\examples{
is_miniconda_installed()
}
\author{
Richèl J.C. Bilderbeek
}
