% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_model.R
\name{is_model}
\alias{is_model}
\title{Determine if the \code{model} is indeed a \code{model}}
\usage{
is_model(model, verbose = FALSE)
}
\arguments{
\item{model}{the \code{GCAE} architecture,
as checked by \link{check_model}.
Use \link{read_model_file} to read a \code{GCAE} architecture from file.
Use \link{save_model} to save the \code{model} to a JSON file.
Use \link{plot_model} to plot the \code{model}.
Use \link{save_model_plot} to save a plot of the \code{model} to file.}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link[plinkr]{check_verbose} to detect if this argument is valid.}
}
\value{
\link{TRUE} if the \code{model} is indeed a \code{model}
}
\description{
Determine if the \code{model} is indeed a \code{model}
}
\examples{
# TRUE
is_model(model = create_test_model())

# FALSE
is_model(model = "nonsense")
is_model(model = "nonsense", verbose = TRUE)
}
\author{
Richèl J.C. Bilderbeek
}
