% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_pip_installed.R
\name{is_pip_installed}
\alias{is_pip_installed}
\title{Determine if \code{pip} is installed}
\usage{
is_pip_installed(gcae_options = create_gcae_options())
}
\arguments{
\item{gcae_options}{options to run GCAE,
as created by \link{create_gcae_options}}
}
\value{
TRUE if \code{pip} is installed
}
\description{
Determine if \code{pip} is installed
}
\examples{
is_pip_installed()
}
\author{
Richèl J.C. Bilderbeek
}
