% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_tensorboard_installed.R
\name{is_tensorboard_installed}
\alias{is_tensorboard_installed}
\title{Determine if TensorBoard is installed}
\usage{
is_tensorboard_installed()
}
\description{
Determine if TensorBoard is installed
}
\author{
Richèl J.C. Bilderbeek
}
