% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_python_packages.R
\name{list_python_packages}
\alias{list_python_packages}
\title{List the Python packages and their versions}
\usage{
list_python_packages(gcae_options = create_gcae_options())
}
\arguments{
\item{gcae_options}{options to run GCAE,
as created by \link{create_gcae_options}}
}
\value{
the Python packages and their versions
}
\description{
List the Python packages and their versions
}
\author{
Richèl J.C. Bilderbeek
}
