% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_true_and_estimated_values.R
\name{normalize_true_and_estimated_values}
\alias{normalize_true_and_estimated_values}
\title{Normalize the true and estimated values}
\usage{
normalize_true_and_estimated_values(true_values, estimated_values)
}
\value{
the result of \link[gganimate]{animate}
}
\description{
Normalize the true and estimated values.
}
\author{
Richèl J.C. Bilderbeek
}
