% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_evaluate_filenames.R
\name{parse_evaluate_filenames}
\alias{parse_evaluate_filenames}
\title{Parse the data from the files produced by \link{gcae_evaluate}}
\usage{
parse_evaluate_filenames(evaluate_filenames, epoch)
}
\arguments{
\item{epoch}{the epoch, as checked by \link{check_epoch}}
}
\value{
a list of tibbles
\itemize{
\item \code{x}: x
}
}
\description{
Parse the data from the files produced by \link{gcae_evaluate}
}
\author{
Richèl J.C. Bilderbeek
}
