% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_genotype_concordances_file.R
\name{parse_genotype_concordances_file}
\alias{parse_genotype_concordances_file}
\title{Parse the \code{genotype_concordances.csv} file}
\usage{
parse_genotype_concordances_file(genotype_concordances_filename)
}
\arguments{
\item{genotype_concordances_filename}{path to the
\code{genotype_concordances.csv} file}
}
\value{
a tibble with columns:
\itemize{
\item \code{n}: the epoch
\item \code{t}: the time this epoch took, in seconds
}
}
\description{
Parse the \code{genotype_concordances.csv} file
}
\author{
Richèl J.C. Bilderbeek
}
