% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_losses_from_project_file.R
\name{parse_losses_from_project_file}
\alias{parse_losses_from_project_file}
\title{Parse the \code{losses_from_project.csv} file}
\usage{
parse_losses_from_project_file(losses_from_project_filename)
}
\arguments{
\item{losses_from_project_filename}{path to the
\code{losses_from_project.csv} file}
}
\value{
a tibble with columns:
\itemize{
\item \code{n}: the epoch
\item \code{t}: the time this epoch took, in seconds
}
}
\description{
Parse the \code{losses_from_project.csv} file
}
\author{
Richèl J.C. Bilderbeek
}
