% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_losses_from_train_v_file.R
\name{parse_losses_from_train_v_file}
\alias{parse_losses_from_train_v_file}
\title{Parse the \code{losses_from_train_v.csv} file}
\usage{
parse_losses_from_train_v_file(losses_from_train_v_filename)
}
\value{
a tibble with columns:
\itemize{
\item \code{n}: the epoch
\item \code{t}: the time this epoch took, in seconds
}
}
\description{
Parse the \code{losses_from_train_v.csv} file
}
\author{
Richèl J.C. Bilderbeek
}
