% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_project_files.R
\name{parse_project_files}
\alias{parse_project_files}
\title{Parse the data from the files produced by \link{gcae_project}}
\usage{
parse_project_files(project_filenames)
}
\arguments{
\item{project_filenames}{path to the files
as created (and returned) by \link{gcae_project}}
}
\value{
a list of these tibbles:
\itemize{
\item \code{losses_from_project_table}: losses from projection
\item \code{genotype_concordances_table}: genotype concordances
}
}
\description{
Parse the data from the files produced by \link{gcae_project}
}
\author{
Richèl J.C. Bilderbeek
}
