% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_score_per_pop_file.R
\name{parse_score_per_pop_file}
\alias{parse_score_per_pop_file}
\title{Parse the \verb{f1_scores_pops_epoch_[x].csv} file
produced by \link{gcae_evaluate}}
\usage{
parse_score_per_pop_file(score_per_pop_filename)
}
\arguments{
\item{score_per_pop_filename}{name of a file containing the evaluation
scores from the dimensionality reduction}
}
\value{
a \link[tibble]{tibble}
}
\description{
Parse the \verb{f1_scores_pops_epoch_[x].csv} file
produced by \link{gcae_evaluate}
}
\examples{
score_per_pop_filename <- get_gcaer_filename("f1_scores_pops_epoch_1.csv")
parse_score_per_pop_file(score_per_pop_filename)
}
\author{
Richèl J.C. Bilderbeek
}
