% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_train_filenames.R
\name{parse_train_filenames}
\alias{parse_train_filenames}
\title{Parse the data from the files produced by \link{gcae_train}}
\usage{
parse_train_filenames(train_filenames)
}
\value{
a list of tibbles
\itemize{
\item \code{train_times}: time in seconds to train each epoch
\item \code{losses_from_train_t}: loss function value on the training set per epoch
\item \code{losses_from_train_v}: loss function value on the validation set per epoch
}
}
\description{
Parse the data from the files produced by \link{gcae_train}
}
\author{
Richèl J.C. Bilderbeek
}
