% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_train_times_file.R
\name{parse_train_times_file}
\alias{parse_train_times_file}
\title{Parse the \code{train_times.csv} file}
\usage{
parse_train_times_file(train_times_filename)
}
\value{
a tibble with columns:
\itemize{
\item \code{n}: the epoch
\item \code{t}: the time this epoch took, in seconds
}
}
\description{
Parse the \code{train_times.csv} file
}
\author{
Richèl J.C. Bilderbeek
}
