% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_genotype_concordances_from_file.R
\name{plot_genotype_concordances_from_file}
\alias{plot_genotype_concordances_from_file}
\title{Plot the genotype concordances in time, from file}
\usage{
plot_genotype_concordances_from_file(
  genotype_concordances_filename,
  png_filename = stringr::str_replace(genotype_concordances_filename, ".csv$", ".png")
)
}
\arguments{
\item{genotype_concordances_filename}{path to the
\code{genotype_concordances.csv} file}

\item{png_filename}{name of the \code{PNG} file to be produced,
as checked by \link{check_png_filename}}
}
\value{
the filename where the plot is stored
}
\description{
Plot the genotype concordances in time, from file
}
\author{
Richèl J.C. Bilderbeek
}
