% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_losses_from_train_t_from_file.R
\name{plot_losses_from_train_t_from_file}
\alias{plot_losses_from_train_t_from_file}
\title{Plot the losses from training in time, for the training set, from file}
\usage{
plot_losses_from_train_t_from_file(
  losses_from_train_t_filename,
  png_filename = stringr::str_replace(losses_from_train_t_filename, ".csv$", ".png")
)
}
\arguments{
\item{losses_from_train_t_filename}{path to the
\code{losses_from_train_t.csv} file}

\item{png_filename}{name of the \code{PNG} file to be produced,
as checked by \link{check_png_filename}}
}
\value{
the filename where the plot is stored
}
\description{
Plot the losses from training in time, for the training set, from file
}
\author{
Richèl J.C. Bilderbeek
}
