% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model.R
\name{plot_model}
\alias{plot_model}
\title{Save a \code{GCAE} autoencoder architecture file}
\usage{
plot_model(model)
}
\arguments{
\item{model}{the \code{GCAE} architecture,
as checked by \link{check_model}.
Use \link{read_model_file} to read a \code{GCAE} architecture from file.
Use \link{save_model} to save the \code{model} to a JSON file.
Use \link{plot_model} to plot the \code{model}.
Use \link{save_model_plot} to save a plot of the \code{model} to file.}
}
\value{
Nothing
}
\description{
Save a \code{GCAE} autoencoder architecture file
}
\seealso{
use \link{read_model_file} to read a \code{model} from file
}
\author{
Richèl J.C. Bilderbeek
}
