% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_phenotype_predictions_from_file.R
\name{plot_phenotype_predictions_from_file}
\alias{plot_phenotype_predictions_from_file}
\title{Plot the phenotype predictions in time, from file}
\usage{
plot_phenotype_predictions_from_file(
  phenotype_predictions_filename,
  png_filename
)
}
\arguments{
\item{phenotype_predictions_filename}{name of a file for
phenotype predictions}

\item{png_filename}{name of the \code{PNG} file to be produced,
as checked by \link{check_png_filename}}
}
\value{
the filename where the plot is stored
}
\description{
Plot the phenotype predictions in time, from file
}
\author{
Richèl J.C. Bilderbeek
}
