% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_r_squared_in_time_from_file.R
\name{plot_r_squared_in_time_from_file}
\alias{plot_r_squared_in_time_from_file}
\title{Plot the \code{r_squared} in time, from file}
\usage{
plot_r_squared_in_time_from_file(r_squared_in_time_filename, png_filename)
}
\arguments{
\item{png_filename}{name of the \code{PNG} file to be produced,
as checked by \link{check_png_filename}}
}
\value{
the filename where the plot is stored
}
\description{
Plot the \code{r_squared} in time, from file
}
\author{
Richèl J.C. Bilderbeek
}
