% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scores_from_file.R
\name{plot_scores_from_file}
\alias{plot_scores_from_file}
\title{Plot the F1 scores in time, from file}
\usage{
plot_scores_from_file(scores_filename, png_filename)
}
\arguments{
\item{scores_filename}{name of the files for F1 scores through time}

\item{png_filename}{name of the \code{PNG} file to be produced,
as checked by \link{check_png_filename}}
}
\value{
the filename where the plot is stored
}
\description{
Plot the F1 scores in time, from file
}
\author{
Richèl J.C. Bilderbeek
}
