% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_train_times_from_file.R
\name{plot_train_times_from_file}
\alias{plot_train_times_from_file}
\title{Plot the train times, from file}
\usage{
plot_train_times_from_file(train_times_filename, png_filename)
}
\arguments{
\item{train_times_filename}{path to the \code{train_times.csv} file}

\item{png_filename}{name of the \code{PNG} file to be produced,
as checked by \link{check_png_filename}}
}
\value{
the filename where the plot is stored
}
\description{
Plot the train times, from file
}
\author{
Richèl J.C. Bilderbeek
}
