% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gcae_experiment_params_file.R
\name{read_gcae_experiment_params_file}
\alias{read_gcae_experiment_params_file}
\title{Read a \code{gcae_experiment_params} from file}
\usage{
read_gcae_experiment_params_file(gcae_experiment_params_filename)
}
\value{
the \code{gcae_experiment_params}
as can be checked by \link{check_gcae_experiment_params}
}
\description{
Read a \code{gcae_experiment_params} from file
}
\examples{
# Create a gcae_experiment_params
gcae_experiment_params <- create_test_gcae_experiment_params()

# Use a temporary file
gcae_experiment_params_filename <- get_gcaer_tempfilename(fileext = ".csv")

# Save gcae_experiment_params to temporary file
save_gcae_experiment_params(
  gcae_experiment_params = gcae_experiment_params,
  gcae_experiment_params_filename = gcae_experiment_params_filename
)

# Load the temporary file to get a copy of the gcae_experiment_params
gcae_experiment_params_again <- read_gcae_experiment_params_file(gcae_experiment_params_filename)

# Cleanup
file.remove(gcae_experiment_params_filename)
}
\seealso{
use \link{save_gcae_experiment_params} to save a \code{gcae_experiment_params}
}
\author{
Richèl J.C. Bilderbeek
}
