% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gcae_input_files.R
\name{read_gcae_input_files}
\alias{read_gcae_input_files}
\title{Read the GenoCAE input filenames}
\usage{
read_gcae_input_files(gcae_input_filenames)
}
\arguments{
\item{gcae_input_filenames}{the GCAE input filenames,
in the form of a \link{list} with the following elements:
\itemize{
\item \code{bed_filename}: path to a \code{.bed} file,
as can be read by \link[plinkr]{read_plink_bed_file}
\item \code{bim_filename}: path to a \code{.bim} file,
as can be read by \link[plinkr]{read_plink_bim_file}
\item \code{fam_filename}: path to a \code{.fam} file,
as can be read by \link[plinkr]{read_plink_fam_file}
\item \code{phe_filename}: path to a \code{.phe} file,
as can be read by \link[plinkr]{read_plink_phe_file}
\item \code{labels_filename}: path to a \code{.csv} file
as can be read by \link{read_labels_file}
}}
}
\value{
the GCAE data in the form of a \link{list}
with the following elements:
\itemize{
\item \code{bed_table}: a \code{.bed} table,
as can be checked by \link[plinkr]{check_bed_table}
\item \code{bim_table}: a \code{.bim} table,
as can be checked by \link[plinkr]{check_bim_table}
\item \code{fam_table}: a \code{.fam} table,
as can be checked by \link[plinkr]{check_fam_table}
\item \code{phe_table}: a \code{.phe} table,
as can be checked by \link[plinkr]{check_phe_table}
\item \code{labels_table}: a \code{.csv} table,
as can be checked by \link{check_labels_table}
}
}
\description{
Read the GenoCAE input filenames
}
\examples{
if (plinkr::is_plink_installed()) {
  # Create the files
  gcae_input_filenames <- create_gcae_input_files_1(
    base_input_filename = get_gcaer_tempfilename()
  )

  # Read the files
  gcae_input_data <- read_gcae_input_files(gcae_input_filenames)

  # Summarise the data
  summarise_gcae_input_data(gcae_input_data)

  # Clean up
  file.remove(as.character(unlist(gcae_input_filenames)))
}
}
\author{
Richèl J.C. Bilderbeek
}
