% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gcae_options_file.R
\name{read_gcae_options_file}
\alias{read_gcae_options_file}
\title{Read a \code{gcae_options} from file}
\usage{
read_gcae_options_file(gcae_options_filename)
}
\arguments{
\item{gcae_options_filename}{name of a file to save/load a \code{gcae_options}
to/from}
}
\value{
the \code{gcae_options}
as can be checked by \link{check_gcae_options}
}
\description{
Read a \code{gcae_options} from file
}
\examples{
# Create a gcae_options
gcae_options <- create_gcae_options()

# Use a temporary file
gcae_options_filename <- get_gcaer_tempfilename(fileext = ".csv")

# Save gcae_options to temporary file
save_gcae_options(
  gcae_options = gcae_options,
  gcae_options_filename = gcae_options_filename
)

# Load the temporary file to get a copy of the gcae_options
gcae_options_again <- read_gcae_options_file(gcae_options_filename)

# Cleanup
file.remove(gcae_options_filename)
}
\seealso{
use \link{save_gcae_options} to save a \code{gcae_options}
}
\author{
Richèl J.C. Bilderbeek
}
