% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gcae_setup_file.R
\name{read_gcae_setup_file}
\alias{read_gcae_setup_file}
\title{Read a \code{gcae_setup} from file}
\usage{
read_gcae_setup_file(gcae_setup_filename)
}
\arguments{
\item{gcae_setup_filename}{name of a file to save/load a \code{gcae_setup}
to/from}
}
\value{
the \code{gcae_setup}
as can be checked by \link{check_gcae_setup}
}
\description{
Read a \code{gcae_setup} from file
}
\examples{
# Create a gcae_setup
gcae_setup <- create_test_gcae_setup()

# Use a temporary file
gcae_setup_filename <- get_gcaer_tempfilename(fileext = ".csv")

# Save gcae_setup to temporary file
save_gcae_setup(
  gcae_setup = gcae_setup,
  gcae_setup_filename = gcae_setup_filename
)

# Load the temporary file to get a copy of the gcae_setup
gcae_setup_again <- read_gcae_setup_file(gcae_setup_filename)

# Cleanup
file.remove(gcae_setup_filename)
}
\seealso{
use \link{save_gcae_setup} to save a \code{gcae_setup}
}
\author{
Richèl J.C. Bilderbeek
}
