% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_genotype_concordances_file.R
\name{read_genotype_concordances_file}
\alias{read_genotype_concordances_file}
\title{Read a \code{genotype_concordances_table} from file}
\usage{
read_genotype_concordances_file(genotype_concordances_filename)
}
\arguments{
\item{genotype_concordances_filename}{path to the
\code{genotype_concordances.csv} file}
}
\value{
the \code{genotype_concordances_table},
as checked by \link{check_genotype_concordances_table}
}
\description{
Read a \code{genotype_concordances_table} from file
}
\author{
Richèl J.C. Bilderbeek
}
