% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_losses_from_train_t_file.R
\name{read_losses_from_train_t_file}
\alias{read_losses_from_train_t_file}
\title{Read a \code{losses_from_train_t_table} from file}
\usage{
read_losses_from_train_t_file(losses_from_train_t_filename)
}
\arguments{
\item{losses_from_train_t_filename}{path to the
\code{losses_from_train_t.csv} file}
}
\value{
the \code{losses_from_train_t_table},
as checked by \link{check_losses_from_train_t_table}
}
\description{
Read a \code{losses_from_train_t_table} from file
}
\author{
Richèl J.C. Bilderbeek
}
