% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_model_file.R
\name{read_model_file}
\alias{read_model_file}
\title{Read a \code{GCAE} autoencoder architecture file}
\usage{
read_model_file(model_filename)
}
\arguments{
\item{model_filename}{name of a file that stores a \code{GCAE} architecture,
as checked by \link{check_model_filename}.
Use \link{read_model_file} to read a \code{GCAE} architecture from that file.}
}
\value{
the \code{model} as a JSON object,
as can be checked by \link{check_model}
}
\description{
Read a \code{GCAE} autoencoder architecture file
}
\examples{
if (is_gcae_repo_cloned()) {
  # A real GCAE file
  model_filename <- get_gcae_model_filename("M1")
} else {
  # An example file
  model_filename <- get_gcaer_filename("M0.json")
}
read_model_file(model_filename)
}
\seealso{
use \link{save_model} to save a \code{model}
}
\author{
Richèl J.C. Bilderbeek
}
