% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nmse_in_time_file.R
\name{read_nmse_in_time_file}
\alias{read_nmse_in_time_file}
\title{Read a \code{nmse_in_time_table} from file}
\usage{
read_nmse_in_time_file(nmse_in_time_filename)
}
\value{
the \code{nmse_in_time_table},
as checked by \link{check_nmse_in_time_table}
}
\description{
Read a \code{nmse_in_time_table} from file
}
\author{
Richèl J.C. Bilderbeek
}
