% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pheno_model_file.R
\name{read_pheno_model_file}
\alias{read_pheno_model_file}
\title{Read a \code{GCAE} phenotypic model file}
\usage{
read_pheno_model_file(pheno_model_filename)
}
\arguments{
\item{pheno_model_filename}{name of a file that stores
the phenotype prediction \code{GCAE} architecture,
as checked by \link{check_pheno_model_filename}.
Use \link{read_pheno_model_file} to read a \code{GCAE}
pehnotype prediction architecture from that file.}
}
\value{
the \code{pheno_model} as a JSON object,
as can be checked by \link{check_pheno_model}
}
\description{
Read a \code{GCAE} phenotypic model file
}
\examples{
if (is_gcae_repo_cloned()) {
  # A real GCAE file
  pheno_model_filename <- get_gcae_pheno_model_filename("p1")
} else {
  # An example file
  pheno_model_filename <- get_gcaer_filename("p0.json")
}
read_pheno_model_file(pheno_model_filename)
}
\seealso{
\itemize{
\item Use \link{save_pheno_model} to save a \code{pheno_model}
\item Use \link{read_model_file} to read a \code{model}
}
}
\author{
Richèl J.C. Bilderbeek
}
