% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_phenotype_predictions_file.R
\name{read_phenotype_predictions_file}
\alias{read_phenotype_predictions_file}
\title{Read a \code{phenotype_predictions_table} from file}
\usage{
read_phenotype_predictions_file(phenotype_predictions_filename)
}
\arguments{
\item{phenotype_predictions_filename}{name of a file for
phenotype predictions}
}
\value{
the \code{phenotype_predictions_table},
as checked by \link{check_phenotype_predictions_table}
}
\description{
Read a \code{phenotype_predictions_table} from file
}
\author{
Richèl J.C. Bilderbeek
}
