% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_r_squared_in_time_file.R
\name{read_r_squared_in_time_file}
\alias{read_r_squared_in_time_file}
\title{Read a \code{r_squared_in_time_table} from file}
\usage{
read_r_squared_in_time_file(r_squared_in_time_filename)
}
\arguments{
\item{r_squared_in_time_filename}{name of the file for
the \code{r_squared} in time.}
}
\value{
the \code{r_squared_in_time_table},
as checked by \link{check_r_squared_in_time_table}
}
\description{
Read a \code{r_squared_in_time_table} from file
}
\author{
Richèl J.C. Bilderbeek
}
