% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_score_per_pop_file.R
\name{read_score_per_pop_file}
\alias{read_score_per_pop_file}
\title{Read a \code{score_per_pop_table} from file}
\usage{
read_score_per_pop_file(score_per_pop_filename)
}
\arguments{
\item{score_per_pop_filename}{name of a file containing the evaluation
scores from the dimensionality reduction}
}
\value{
the \code{score_per_pop_table},
as checked by \link{check_score_per_pop_table}
}
\description{
Read a \code{score_per_pop_table} from file
}
\author{
Richèl J.C. Bilderbeek
}
