% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_scores_file.R
\name{read_scores_file}
\alias{read_scores_file}
\title{Read a \code{scores_table} from file}
\usage{
read_scores_file(scores_filename)
}
\value{
the \code{scores_table},
as checked by \link{check_scores_table}
}
\description{
Read a \code{scores_table} from file
}
\author{
Richèl J.C. Bilderbeek
}
