% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_scores_per_pop_file.R
\name{read_scores_per_pop_file}
\alias{read_scores_per_pop_file}
\title{Read a \code{scores_per_pop_table} from file}
\usage{
read_scores_per_pop_file(scores_per_pop_filename)
}
\value{
the \code{scores_per_pop_table},
as checked by \link{check_scores_per_pop_table}
}
\description{
Read a \code{scores_per_pop_table} from file
}
\author{
Richèl J.C. Bilderbeek
}
