% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_train_times_file.R
\name{read_train_times_file}
\alias{read_train_times_file}
\title{Read a \code{train_times_table} from file}
\usage{
read_train_times_file(train_times_filename)
}
\arguments{
\item{train_times_filename}{path to the \code{train_times.csv} file}
}
\value{
the \code{train_times_table},
as checked by \link{check_train_times_table}
}
\description{
Read a \code{train_times_table} from file
}
\author{
Richèl J.C. Bilderbeek
}
