% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_gcaer_tempfolder.R
\name{remove_gcaer_tempfolder}
\alias{remove_gcaer_tempfolder}
\title{Remove the \code{gcaer} temporary folder}
\usage{
remove_gcaer_tempfolder()
}
\value{
Nothing
}
\description{
Remove the \code{gcaer} temporary folder
}
\examples{
if (plinkr::is_on_ci()) {
  remove_gcaer_tempfolder()
}
}
\author{
Richèl J.C. Bilderbeek
}
