% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resize_to_shared_individuals_from_data.R
\name{resize_to_shared_individuals_from_data}
\alias{resize_to_shared_individuals_from_data}
\title{Resize the \code{gcae_input_data} to only contain the shared individuals.}
\usage{
resize_to_shared_individuals_from_data(gcae_input_data, verbose = FALSE)
}
\arguments{
\item{gcae_input_data}{in-memory data that can be used
as input for \code{GCAE}. Use link{read_gcae_input_files}
to read it from files, use \link{check_gcae_input_data}
to check it for validity.}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link[plinkr]{check_verbose} to detect if this argument is valid.}
}
\value{
the \code{gcae_input_data} with only shared individuals
}
\description{
Resize the \code{gcae_input_data} to only contain the shared individuals.
}
\examples{
gcae_input_data <- create_test_gcae_input_data()

# Show before
summarise_gcae_input_data(gcae_input_data)

gcae_input_data <- resize_to_shared_individuals_from_data(gcae_input_data)

# Show after
summarise_gcae_input_data(gcae_input_data)
}
\seealso{
Related checking functions:
\itemize{
\item Use \link{check_gcae_input_data} to check the in-memory GCAE input data
extensively
\item Use \link{check_gcae_input_data_data_type} to check the
data type of the in-memory GCAE input data
\item Use \link{check_gcae_input_files} to check the
files to be used by GCAE as input,
i.e. \code{GCAE} can run these
\item Use \link{check_gcae_input_filenames} to check the
filenames (and not the files' content) to be GCAE input files
}
}
\author{
Richèl J.C. Bilderbeek
}
