% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resize_to_shared_individuals_from_files.R
\name{resize_to_shared_individuals_from_files}
\alias{resize_to_shared_individuals_from_files}
\title{Resize the \code{gcae_input_data} to only contain the shared individuals.}
\usage{
resize_to_shared_individuals_from_files(gcae_input_filenames)
}
\arguments{
\item{gcae_input_filenames}{the GCAE input filenames,
in the form of a \link{list} with the following elements:
\itemize{
\item \code{bed_filename}: path to a \code{.bed} file,
as can be read by \link[plinkr]{read_plink_bed_file}
\item \code{bim_filename}: path to a \code{.bim} file,
as can be read by \link[plinkr]{read_plink_bim_file}
\item \code{fam_filename}: path to a \code{.fam} file,
as can be read by \link[plinkr]{read_plink_fam_file}
\item \code{phe_filename}: path to a \code{.phe} file,
as can be read by \link[plinkr]{read_plink_phe_file}
\item \code{labels_filename}: path to a \code{.csv} file
as can be read by \link{read_labels_file}
}}
}
\value{
the \code{gcae_input_data} with only shared individuals
}
\description{
Resize the \code{gcae_input_data} to only contain the shared individuals.
}
\examples{
gcae_input_data <- create_test_gcae_input_data()

# Show before
summarise_gcae_input_data(gcae_input_data)

gcae_input_data <- resize_to_shared_individuals_from_data(gcae_input_data)

# Show after
summarise_gcae_input_data(gcae_input_data)
}
\seealso{
Related checking functions:
\itemize{
\item Use \link{check_gcae_input_data} to check the in-memory GCAE input data
extensively
\item Use \link{check_gcae_input_data_data_type} to check the
data type of the in-memory GCAE input data
\item Use \link{check_gcae_input_files} to check the
files to be used by GCAE as input,
i.e. \code{GCAE} can run these
\item Use \link{check_gcae_input_filenames} to check the
filenames (and not the files' content) to be GCAE input files
}
}
\author{
Richèl J.C. Bilderbeek
}
