% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_gcae.R
\name{run_gcae}
\alias{run_gcae}
\title{Run GCAE from arguments}
\usage{
run_gcae(args, gcae_options = create_gcae_options(), verbose = FALSE)
}
\arguments{
\item{args}{arguments for the GCAE executable}

\item{gcae_options}{options to run GCAE,
as created by \link{create_gcae_options}}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link[plinkr]{check_verbose} to detect if this argument is valid.}
}
\value{
the text that GCAE returns
}
\description{
Run GCAE from arguments.
In some contexts, e.g. \link{gcae_train},
the working directory must be the same as the folder the
\code{run_gcae.py} scipt is in (see
\url{https://github.com/kausmees/GenoCAE/issues/10})
}
\examples{
if (is_gcae_installed()) {
  run_gcae("--help")
}
}
\author{
Richèl J.C. Bilderbeek
}
