% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_gcae_experiment_params.R
\name{save_gcae_experiment_params}
\alias{save_gcae_experiment_params}
\title{Save a \code{gcae_experiment_params} to file}
\usage{
save_gcae_experiment_params(
  gcae_experiment_params,
  gcae_experiment_params_filename
)
}
\arguments{
\item{gcae_experiment_params}{parameters to run a full
\code{GCAE} experimenr (i.e. call \link{do_gcae_experiment}),
as created by \link{create_gcae_experiment_params} and checked
by \link{check_gcae_experiment_params}}
}
\value{
Nothing
}
\description{
Save a \code{gcae_experiment_params} to file
}
\examples{
# Create a gcae_experiment_params
gcae_experiment_params <- create_test_gcae_experiment_params()

# Use a temporary file
gcae_experiment_params_filename <- get_gcaer_tempfilename(fileext = ".csv")

# Save gcae_experiment_params to temporary file
save_gcae_experiment_params(
  gcae_experiment_params = gcae_experiment_params,
  gcae_experiment_params_filename = gcae_experiment_params_filename
)

# Load the temporary file to get a copy of the gcae_experiment_params
gcae_experiment_params_again <- read_gcae_experiment_params_file(gcae_experiment_params_filename)

# Cleanup
file.remove(gcae_experiment_params_filename)
}
\seealso{
Use \link{read_gcae_experiment_params_file} to read
a \code{gcae_experiment_params} from file
}
\author{
Richèl J.C. Bilderbeek
}
