% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_gcae_experiment_results.R
\name{save_gcae_experiment_results}
\alias{save_gcae_experiment_results}
\title{Save a \code{gcae_experiment_results} to file}
\usage{
save_gcae_experiment_results(gcae_experiment_results, gcae_experiment_params)
}
\arguments{
\item{gcae_experiment_results}{the results of a call
to \link{do_gcae_experiment},
as checked by \link{check_gcae_experiment_results}}

\item{gcae_experiment_params}{parameters to run a full
\code{GCAE} experimenr (i.e. call \link{do_gcae_experiment}),
as created by \link{create_gcae_experiment_params} and checked
by \link{check_gcae_experiment_params}}
}
\value{
Nothing
}
\description{
Save a \code{gcae_experiment_results} to file
}
\examples{
# Create a gcae_experiment_results
gcae_experiment_results <- create_test_gcae_experiment_results()

# Use a temporary file
gcae_experiment_results_filename <- get_gcaer_tempfilename(fileext = ".csv")

# Save gcae_experiment_results to temporary file
save_gcae_experiment_results(
  gcae_experiment_results = gcae_experiment_results,
  gcae_experiment_results_filename = gcae_experiment_results_filename
)

# Load the temporary file to get a copy of the gcae_experiment_results
gcae_experiment_results_again <- read_gcae_experiment_results_file(
  gcae_experiment_results_filename
)

# Cleanup
file.remove(gcae_experiment_results_filename)
}
\seealso{
Use \link{read_gcae_experiment_results_file} to read
a \code{gcae_experiment_results} from file
}
\author{
Richèl J.C. Bilderbeek
}
