% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_gcae_input_data.R
\name{save_gcae_input_data}
\alias{save_gcae_input_data}
\title{Save a \code{gcae_input_data}}
\usage{
save_gcae_input_data(gcae_input_data, gcae_input_filenames)
}
\arguments{
\item{gcae_input_data}{in-memory data that can be used
as input for \code{GCAE}. Use link{read_gcae_input_files}
to read it from files, use \link{check_gcae_input_data}
to check it for validity.}

\item{gcae_input_filenames}{the GCAE input filenames,
in the form of a \link{list} with the following elements:
\itemize{
\item \code{bed_filename}: path to a \code{.bed} file,
as can be read by \link[plinkr]{read_plink_bed_file}
\item \code{bim_filename}: path to a \code{.bim} file,
as can be read by \link[plinkr]{read_plink_bim_file}
\item \code{fam_filename}: path to a \code{.fam} file,
as can be read by \link[plinkr]{read_plink_fam_file}
\item \code{phe_filename}: path to a \code{.phe} file,
as can be read by \link[plinkr]{read_plink_phe_file}
\item \code{labels_filename}: path to a \code{.csv} file
as can be read by \link{read_labels_file}
}}
}
\value{
the \code{gcae_input_data} invisibly
}
\description{
Save a \code{gcae_input_data}
}
\author{
Richèl J.C. Bilderbeek
}
