% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_gcae_options.R
\name{save_gcae_options}
\alias{save_gcae_options}
\title{Save a \code{gcae_options} to file}
\usage{
save_gcae_options(
  gcae_options = gcae_options,
  gcae_options_filename = gcae_options_filename
)
}
\arguments{
\item{gcae_options}{options to run GCAE,
as created by \link{create_gcae_options}}

\item{gcae_options_filename}{name of a file to save/load a \code{gcae_options}
to/from}
}
\value{
Nothing
}
\description{
Save a \code{gcae_options} to file
}
\examples{
# Create a gcae_options
gcae_options <- create_gcae_options()

# Use a temporary file
gcae_options_filename <- get_gcaer_tempfilename(fileext = ".csv")

# Save gcae_options to temporary file
save_gcae_options(
  gcae_options = gcae_options,
  gcae_options_filename = gcae_options_filename
)

# Load the temporary file to get a copy of the gcae_options
gcae_options_again <- read_gcae_options_file(gcae_options_filename)

# Cleanup
file.remove(gcae_options_filename)
}
\seealso{
Use \link{read_gcae_options_file} to read a \code{gcae_options} from file
}
\author{
Richèl J.C. Bilderbeek
}
