% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_gcae_setup.R
\name{save_gcae_setup}
\alias{save_gcae_setup}
\title{Save a \code{gcae_setup} to file}
\usage{
save_gcae_setup(
  gcae_setup = gcae_setup,
  gcae_setup_filename = gcae_setup_filename
)
}
\arguments{
\item{gcae_setup}{GCAE model setup,
as created by \link{create_gcae_setup},
read from file by \link{read_gcae_setup_file},
checked by \link{check_gcae_setup},
and saved to file by \link{save_gcae_setup}.}

\item{gcae_setup_filename}{name of a file to save/load a \code{gcae_setup}
to/from}
}
\value{
Nothing
}
\description{
Save a \code{gcae_setup} to file
}
\examples{
# Create a gcae_setup
gcae_setup <- create_test_gcae_setup()

# Use a temporary file
gcae_setup_filename <- get_gcaer_tempfilename(fileext = ".csv")

# Save gcae_setup to temporary file
save_gcae_setup(
  gcae_setup = gcae_setup,
  gcae_setup_filename = gcae_setup_filename
)

# Load the temporary file to get a copy of the gcae_setup
gcae_setup_again <- read_gcae_setup_file(gcae_setup_filename)

# Cleanup
file.remove(gcae_setup_filename)
}
\seealso{
Use \link{read_gcae_setup_file} to read a \code{gcae_setup} from file
}
\author{
Richèl J.C. Bilderbeek
}
