% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_labels_file.R
\name{save_labels_file}
\alias{save_labels_file}
\title{Save a \code{GCAE} labels file}
\usage{
save_labels_file(labels_table, labels_filename)
}
\arguments{
\item{labels_table}{a table containing super-
and -population labels, as can be read by \link{read_labels_table}.
Note that \code{GCAE} does not use column names, \code{gcaer} adds these
for convenience}

\item{labels_filename}{path to a \code{.csv} file containing super-
and -population labels, as can be read by \link{read_labels_table}}
}
\value{
the \code{labels_table} invisibly
}
\description{
Save a \code{GCAE} labels file
}
\note{
\code{GCAE} does not save this data with column names
}
\examples{
# Get a labels table
labels_filename <- get_gcaer_filename("gcae_input_files_1_labels.csv")
labels_table <- read_labels_file(labels_filename)

# Save it
labels_filename <- get_gcaer_tempfilename(fileext = ".csv")
save_labels_file(
  labels_table = labels_table,
  labels_filename = labels_filename
)

# Clean up
file.remove(labels_filename)
}
\seealso{
use \link{read_labels_file} to load a \code{labels_table}
from file
}
\author{
Richèl J.C. Bilderbeek
}
