% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_model.R
\name{save_model}
\alias{save_model}
\title{Save a \code{GCAE} autoencoder architecture file}
\usage{
save_model(model, model_filename)
}
\arguments{
\item{model}{the \code{GCAE} architecture,
as checked by \link{check_model}.
Use \link{read_model_file} to read a \code{GCAE} architecture from file.}

\item{model_filename}{name of a file that stores a \code{GCAE} architecture,
as checked by \link{check_model_filename}.
Use \link{read_model_file} to read a \code{GCAE} architecture from that file.}
}
\value{
Nothing
}
\description{
Save a \code{GCAE} autoencoder architecture file
}
\seealso{
use \link{read_model_file} to read a \code{model} from file
}
\author{
Richèl J.C. Bilderbeek
}
