% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_model_plot.R
\name{save_model_plot}
\alias{save_model_plot}
\title{Save a \code{GCAE} autoencoder architecture file}
\usage{
save_model_plot(model, png_filename)
}
\arguments{
\item{model}{the \code{GCAE} architecture,
as checked by \link{check_model}.
Use \link{read_model_file} to read a \code{GCAE} architecture from file.
Use \link{save_model} to save the \code{model} to a JSON file.
Use \link{plot_model} to plot the \code{model}.
Use \link{save_model_plot} to save a plot of the \code{model} to file.}

\item{png_filename}{name of the \code{PNG} file to be produced,
as checked by \link{check_png_filename}}
}
\value{
Nothing
}
\description{
Save a \code{GCAE} autoencoder architecture file
}
\seealso{
use \link{save_model} to read the \code{model} to a JSON file
}
\author{
Richèl J.C. Bilderbeek
}
