% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_gcae_input_files.R
\name{summarise_gcae_input_files}
\alias{summarise_gcae_input_files}
\title{Summarise the GCAE input files}
\usage{
summarise_gcae_input_files(gcae_input_filenames, verbose = FALSE)
}
\arguments{
\item{gcae_input_filenames}{the GCAE input filenames,
in the form of a \link{list} with the following elements:
\itemize{
\item \code{bed_filename}: path to a \code{.bed} file,
as can be read by \link[plinkr]{read_plink_bed_file}
\item \code{bim_filename}: path to a \code{.bim} file,
as can be read by \link[plinkr]{read_plink_bim_file}
\item \code{fam_filename}: path to a \code{.fam} file,
as can be read by \link[plinkr]{read_plink_fam_file}
\item \code{phe_filename}: path to a \code{.phe} file,
as can be read by \link[plinkr]{read_plink_phe_file}
\item \code{labels_filename}: path to a \code{.csv} file
as can be read by \link{read_labels_file}
}}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link[plinkr]{check_verbose} to detect if this argument is valid.}
}
\value{
This function will show the summary using \link{message}.
A \link{list} with all the elements shown is returned as well
}
\description{
Summarise the GCAE input files
}
\note{
this function is named after the base R \link{summarise} function
}
\examples{
gcae_input_filenames <- create_test_gcae_input_filenames()
summarise_gcae_input_files(gcae_input_filenames)
}
\author{
Richèl J.C. Bilderbeek
}
