% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uninstall_gcae.R
\name{uninstall_gcae}
\alias{uninstall_gcae}
\title{Uninstall GCAE}
\usage{
uninstall_gcae(gcae_options = create_gcae_options())
}
\arguments{
\item{gcae_options}{options to run GCAE,
as created by \link{create_gcae_options}}
}
\value{
Nothing
}
\description{
Uninstall GCAE
}
\author{
Richèl J.C. Bilderbeek
}
