% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upgrade_pip.R
\name{upgrade_pip}
\alias{upgrade_pip}
\title{Upgrade pip}
\usage{
upgrade_pip(verbose = FALSE)
}
\arguments{
\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link[plinkr]{check_verbose} to detect if this argument is valid.}
}
\value{
the text of the installation
}
\description{
Upgrade pip
}
