% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_qt_prediction.R
\name{analyse_qt_prediction}
\alias{analyse_qt_prediction}
\title{Determine how well the quantitative trait prediction worked}
\usage{
analyse_qt_prediction(
  datadir,
  trainedmodeldir,
  png_filename,
  csv_filename_for_mse,
  csv_filename_for_fits,
  csv_filename_for_r_squareds,
  verbose = FALSE
)
}
\arguments{
\item{datadir}{directory where sample data is stored.
If not absolute: assumed relative to \verb{GenoCAE/} directory.
A valid \code{datadir} ends with a slash (\code{/})
and does not start with a full stop (\code{.}),
as to mimic the \code{GCAE} interface}

\item{trainedmodeldir}{name of the folder to save the trained model to.
A valid \code{trainedmodeldir} ends with a slash (\code{/})
as to mimic the \code{GCAE} interface}

\item{png_filename}{name of the \code{PNG} file to be produced,
as checked by \link{check_png_filename}}

\item{csv_filename_for_mse}{name for the comma-seperated file
(as checked by \link{check_csv_filename})
to store the mean square error compared to the identity line to}

\item{csv_filename_for_fits}{name for the comma-seperated file
(as checked by \link{check_csv_filename})
to store the fits of linear, quadratic and cubic models to}

\item{csv_filename_for_r_squareds}{name for the comma-seperated file
(as checked by \link{check_csv_filename})
to store the multiple values of R squared
(i.e. of linear, quadratic and cubic models) to}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link[plinkr]{check_verbose} to detect if this argument is valid.}
}
\value{
the plot
}
\description{
Determine how well the quantitative trait prediction worked
}
\author{
Richèl J.C. Bilderbeek
}
