% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_mse_from_identity_line.R
\name{calc_mse_from_identity_line}
\alias{calc_mse_from_identity_line}
\title{Calculate the mean squared error from the identity line.}
\usage{
calc_mse_from_identity_line(true_values, estimated_values)
}
\arguments{
\item{true_values}{the true values}

\item{estimated_values}{the estimated values}
}
\value{
the mean squared error from the identity line
}
\description{
Calculate the mean squared error from the identity line
(i.e. the line \code{y = x}).
If the estimated values are equal to the true values, the MSE is zero.
}
\examples{
true_values_degrees <- seq(1, 360)
estimated_values_degrees <- true_values_degrees + stats::runif(n = 360)
mse_degrees <- calc_mse_from_identity_line(
  true_values = true_values_degrees,
  estimated_values = estimated_values_degrees
)
mse_degrees

# Convert degrees to radians
true_values_radians <- 2.0 * pi * true_values_degrees / 360
estimated_values_radians <- 2.0 * pi * estimated_values_degrees / 360
mse_radians <- calc_mse_from_identity_line(
  true_values = true_values_radians,
  estimated_values = estimated_values_radians
)
mse_radians

# mse_degrees and mse_radians differ, due to their different scales,
# use 'calc_nmse_from_identity_line' to normalize the scale first
}
\seealso{
use \link{calc_nmse_from_identity_line} to calculate the
normalized mean square error from the identity line
}
\author{
Richèl J.C. Bilderbeek
}
