% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{check_data}
\alias{check_data}
\title{Check if the \code{data} is valid}
\usage{
check_data(data)
}
\arguments{
\item{data}{file prefix, not including path, of the data files
(EIGENSTRAT of PLINK format)}
}
\value{
Nothing. Will \link{stop} if \code{data} is invalid.
}
\description{
Check if the \code{data} is valid
Will \link{stop} if not.
}
\author{
Richèl J.C. Bilderbeek
}
