% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_epoch.R
\name{check_epoch}
\alias{check_epoch}
\title{Check if the \code{epoch} is valid}
\usage{
check_epoch(epoch)
}
\arguments{
\item{epoch}{the epoch, as checked by \link{check_epoch}}
}
\value{
Nothing. Will \link{stop} if \code{epoch} is invalid.
}
\description{
Check if the \code{epoch} is valid
Will \link{stop} if not.
}
\examples{
check_epoch(100)
}
\author{
Richèl J.C. Bilderbeek
}
