% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_model.R
\name{check_model}
\alias{check_model}
\title{Check if the \code{model} is valid}
\usage{
check_model(model)
}
\arguments{
\item{model}{the \code{GCAE} architecture,
as checked by \link{check_model}.
Use \link{read_model_file} to read a \code{GCAE} architecture from file.}
}
\value{
Nothing. Will \link{stop} if \code{model} is invalid.
}
\description{
Check if the \code{model} is valid
Will \link{stop} if not.
}
\examples{
model_filename <- get_gcae_model_filename("M0")
model <- read_model_file(model_filename)
check_model(model)

model_filename <- get_gcae_model_filename("M1")
model <- read_model_file(model_filename)
check_model(model)
}
\author{
Richèl J.C. Bilderbeek
}
