% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_model_filename.R
\name{check_model_filename}
\alias{check_model_filename}
\title{Check if the \code{model_filename} is valid}
\usage{
check_model_filename(model_filename)
}
\value{
Nothing. Will \link{stop} if \code{model_filename} is invalid.
}
\description{
Check if the \code{model_filename} is valid
Will \link{stop} if not.
}
\examples{
check_model_filename("M0.json")
check_model_filename("~/M0.json")
check_model_filename("/home/me/M2.json")
}
\author{
Richèl J.C. Bilderbeek
}
