% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_gcae_experiment_params.R
\name{create_gcae_experiment_params}
\alias{create_gcae_experiment_params}
\title{Create the GCAE setup}
\usage{
create_gcae_experiment_params(gcae_options, gcae_setup, analyse_epochs)
}
\arguments{
\item{gcae_options}{options to run GCAE,
as created by \link{create_gcae_options}}

\item{gcae_setup}{GCAE model setup,
as created by \link{create_gcae_setup},
read from file by \link{read_gcae_setup_file},
checked by \link{check_gcae_setup},
and saved to file by \link{save_gcae_setup}.}
}
\value{
full paths to the created files
}
\description{
Create the GCAE setup
}
\examples{
create_gcae_experiment_params(
  gcae_options = create_gcae_options(),
  gcae_setup = create_test_gcae_setup(),
  analyse_epochs = seq(1, 2, 3)
)
}
\author{
Richèl J.C. Bilderbeek
}
