% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_gcae_setup.R
\name{create_gcae_setup}
\alias{create_gcae_setup}
\title{Create the GCAE setup}
\usage{
create_gcae_setup(
  model_id = "M1",
  train_opts_id = "ex3",
  data_opts_id = "b_0_4"
)
}
\arguments{
\item{model_id}{model id, corresponding to a file models/model_id.json}

\item{train_opts_id}{train options id, corresponding to a file
train_opts/train_opts_id.json}

\item{data_opts_id}{data options id,
corresponding to a file data_opts/data_opts_id.json}
}
\value{
full paths to the created files
}
\description{
Create the GCAE setup, which is:
\itemize{
\item \code{model} setup of the autoencoder
\item \code{train_options} setup of the training
\item \code{data_opts_id} setup of ?
}
}
\author{
Richèl J.C. Bilderbeek
}
