% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_gcae_train_args.R
\name{create_gcae_train_args}
\alias{create_gcae_train_args}
\title{Create the CLI arguments to train with GCAE}
\usage{
create_gcae_train_args(gcae_setup, epochs, save_interval)
}
\arguments{
\item{gcae_setup}{GCAE model setup,
as created by \link{create_gcae_setup}}

\item{epochs}{number of epochs to train}

\item{save_interval}{epoch intervals at which to save state of model,
and at which to calculate the valid loss}
}
\value{
the CLI arguments
}
\description{
Create the CLI arguments to train with GCAE
}
\examples{
gcae_setup <- create_gcae_setup(
  datadir = file.path(get_gcae_folder(), "example_tiny/"),
  data = "issue_6_bin",
  model_id = "M1",
  pheno_model_id = "p2"
)
args <- create_gcae_train_args(
  gcae_setup = gcae_setup,
  epochs = 1,
  save_interval = 1
)
# 'python3' is the path to Python3
# 'run_gcae.py' is the path to the GenoCAE Python script
full_args <- c("python3", "run_gcae.py", args)
cat(full_args)
}
\author{
Richèl J.C. Bilderbeek
}
