% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_test_gcae_input_data.R
\name{create_test_gcae_input_data}
\alias{create_test_gcae_input_data}
\title{Create a \code{gcae_input_data} to be used in testing.}
\usage{
create_test_gcae_input_data(
  gcae_input_filenames = gcaer::create_test_gcae_input_filenames(),
  verbose = FALSE
)
}
\arguments{
\item{gcae_input_filenames}{the GCAE input filenames,
in the form of a \link{list} with the following elements:
\itemize{
\item \code{bed_filename}: path to a \code{.bed} file,
as can be read by \link[plinkr]{read_plink_bed_file}
\item \code{bim_filename}: path to a \code{.bim} file,
as can be read by \link[plinkr]{read_plink_bim_file}
\item \code{fam_filename}: path to a \code{.fam} file,
as can be read by \link[plinkr]{read_plink_fam_file}
\item \code{phe_filename}: path to a \code{.phe} file,
as can be read by \link[plinkr]{read_plink_phe_file}
\item \code{labels_filename}: path to a \code{.csv} file
as can be read by \link{read_labels_file}
}}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link[plinkr]{check_verbose} to detect if this argument is valid.}
}
\value{
a \code{gcae_input_data}
}
\description{
Create a \code{gcae_input_data} to be used in testing.
}
\examples{
gcae_input_data <- create_test_gcae_input_data()
check_gcae_input_data(gcae_input_data)
}
\seealso{
use \link{create_gcae_input_data} to create
a regular \code{gcae_input_data}
}
\author{
Richèl J.C. Bilderbeek
}
