% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_params_doc.R
\name{default_params_doc}
\alias{default_params_doc}
\title{This function does nothing. It is intended to inherit is parameters'
documentation.}
\usage{
default_params_doc(
  args,
  base_input_filename,
  base_output_filename,
  base_phenotype_value,
  data,
  datadir,
  data_opts_id,
  epochs,
  example_filename,
  gcae_exe_path,
  gcae_folder,
  gcae_input_filenames,
  gcae_options,
  gcae_setup,
  gcae_version,
  gcaer_folder,
  genotype_concordances_filename,
  github_repo_branch_name,
  github_repo_url,
  labels_filename,
  labels_table,
  log_filename,
  losses_from_project_filename,
  losses_from_train_t_filename,
  losses_from_train_v_filename,
  model_id,
  ormr_folder_name,
  os,
  out,
  pheno_model_id,
  project_filenames,
  python_bin_path,
  python_version,
  save_interval,
  superpops,
  train_filenames,
  train_opts_id,
  train_times_filename,
  trainedmodeldir,
  trainedmodelname,
  url,
  verbose
)
}
\arguments{
\item{args}{arguments for the \code{GCAE} executable}

\item{base_input_filename}{the base of the filenames that are
used as input for \code{GCAE}/\code{GCAE2}}

\item{base_output_filename}{the base of the filenames that are
used as output for \code{GCAE}/\code{GCAE2}}

\item{base_phenotype_value}{the base phenotypic value for an additive trait,
i.e. the phenotypic value for homozygotes of the common allele}

\item{data}{file prefix, not including path, of the data files.
The data files must be in EIGENSTRAT
or PLINK binary (\code{.bed}, \code{.bim}, \code{.fam}) format)}

\item{datadir}{directory where sample data is stored.
If not absolute: assumed relative to \verb{GenoCAE/} directory.
A valid \code{datadir} ends with a slash (\code{/})
and does not start with a full stop (\code{.})}

\item{data_opts_id}{data options id,
corresponding to a file data_opts/data_opts_id.json
The data options affect how data is input to the model.}

\item{epochs}{number of epochs to train}

\item{example_filename}{name of the example file}

\item{gcae_exe_path}{path to
the \code{GCAE} or \code{GCAE2} executable file.}

\item{gcae_folder}{folder where \code{GCAE} is installed}

\item{gcae_input_filenames}{the GCAE input filenames,
in the form of a \link{list} with the following elements:
\itemize{
\item \code{bed_filename}: path to a \code{.bed} file,
as can be read by \link[plinkr]{read_plink_bed_file}
\item \code{bim_filename}: path to a \code{.bim} file,
as can be read by \link[plinkr]{read_plink_bim_file}
\item \code{fam_filename}: path to a \code{.fam} file,
as can be read by \link[plinkr]{read_plink_fam_file}
\item \code{phe_filename}: path to a \code{.phe} file,
as can be read by \link[plinkr]{read_plink_phe_file}
\item \code{labels_filename}: path to a \code{.csv} file
as can be read by \link{read_labels_file}
}}

\item{gcae_options}{options to run GCAE,
as created by \link{create_gcae_options}}

\item{gcae_setup}{GCAE model setup,
as created by \link{create_gcae_setup}}

\item{gcae_version}{version of GCAE, e.g. \code{"1.0"}
Use \link{get_gcae_version} to get the \code{GCAE} version.}

\item{gcaer_folder}{name of the folder where \link{gcaer}
stores its temporary files}

\item{genotype_concordances_filename}{path to the
\code{genotype_concordances.csv} file}

\item{github_repo_branch_name}{name of the \code{git} branch,
e.g \code{master} or \code{Pheno}.
Use \link{get_default_github_repo_branch_name} for the default value}

\item{github_repo_url}{URL of the GitHub repository,
e.g. \verb{https://github.com/kausmees/GenoCAE}
or \verb{https://github.com/cnettel/GenoCAE}
or \verb{https://github.com/richelbilderbeek/GenoCAE}.
Use \link{get_default_github_repo_url} for the default value}

\item{labels_filename}{path to a \code{.csv} file containing super-
and -population labels, as can be read by \link{read_labels_table}}

\item{labels_table}{a table containing super-
and -population labels, as can be read by \link{read_labels_table}.
Note that \code{GCAE} does not use column names, \code{gcaer} adds these
for convenience}

\item{log_filename}{name of a \code{GCAE} \code{.log} file}

\item{losses_from_project_filename}{path to the
\code{losses_from_project.csv} file}

\item{losses_from_train_t_filename}{path to the
\code{losses_from_train_t.csv} file}

\item{losses_from_train_v_filename}{path to the
\code{losses_from_train_v.csv} file}

\item{model_id}{model id, corresponding to a file models/model_id.json}

\item{ormr_folder_name}{folder where the link[ormr]{ormr} R package
installs the Python packages}

\item{os}{name of the operating system,
as returned by \link[rappdirs]{app_dir}}

\item{out}{the base filename of the output files.
This parameter is named after the \code{GCAE}
\code{--out} flag}

\item{pheno_model_id}{phenotype model ID,
as checked by \link{check_pheno_model_id}}

\item{project_filenames}{path to the files
as created (and returned) by \link{gcae_project}}

\item{python_bin_path}{path to a \code{python} binary/executable}

\item{python_version}{the Python version}

\item{save_interval}{epoch intervals at which to save state of model,
and at which to calculate the valid loss}

\item{superpops}{path to the \code{superpops} file, which contains
the labels table and can be read using \link{read_labels_file}}

\item{train_filenames}{path to the files
as created (and returned) by \link{gcae_train}}

\item{train_opts_id}{train options id, corresponding to a file
train_opts/train_opts_id.json}

\item{train_times_filename}{path to the \code{train_times.csv} file}

\item{trainedmodeldir}{name of the folder to save the trained model to.}

\item{trainedmodelname}{name of the trained model,
which is \code{ae_out} by default}

\item{url}{\code{GCAE} download URL}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link[plinkr]{check_verbose} to detect if this argument is valid.}
}
\description{
This function does nothing. It is intended to inherit is parameters'
documentation.
}
\note{
This is an internal function, so it should be marked with
\code{@noRd}. This is not done, as this will disallow all
functions to find the documentation parameters
}
\author{
Richèl J.C. Bilderbeek
}
