% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_params_doc.R
\name{default_params_doc}
\alias{default_params_doc}
\title{This function does nothing. It is intended to inherit is parameters'
documentation.}
\usage{
default_params_doc(
  analyse_epochs,
  args,
  base_input_filename,
  base_output_filename,
  base_phenotype_value,
  bed_filename,
  bim_filename,
  csv_filename,
  csv_filename_for_nmse,
  data,
  datadir,
  data_opts_id,
  epoch,
  epochs,
  estimated_values,
  evaluate_filenames,
  evaluate_results,
  example_filename,
  fam_filename,
  file_extension,
  folder_name,
  gcae_experiment_params,
  gcae_experiment_params_filename,
  gcae_experiment_results,
  gcae_folder,
  gcae_input_data,
  gcae_input_filenames,
  gcae_options,
  gcae_options_filename,
  gcae_setup,
  gcae_setup_filename,
  gcae_version,
  gcaer_folder,
  genotype_concordances_filename,
  genotype_concordances_table,
  gif_filename,
  github_repo_branch_name,
  github_repo_url,
  labels_filename,
  labels_table,
  log_filename,
  losses_from_project_filename,
  losses_from_train_t_filename,
  losses_from_train_t_table,
  losses_from_train_v_filename,
  losses_from_train_v_table,
  metric,
  metrics,
  model,
  model_filename,
  model_id,
  n_neurons,
  nmse_in_time_filename,
  nmse_in_time_table,
  ormr_folder_name,
  os,
  out,
  phe_filename,
  plink_options,
  pheno_model_filename,
  pheno_model_id,
  phenotype_predictions_filename,
  phenotype_predictions_table,
  png_filename,
  project_filenames,
  python_bin_path,
  python_version,
  resume_from,
  save_interval,
  score_per_pop_filename,
  score_per_pop_table,
  scores_filename,
  scores_table,
  superpops,
  train_filenames,
  train_times_table,
  train_opts_id,
  train_times_filename,
  trainedmodeldir,
  trainedmodelname,
  true_values,
  unique_id,
  url,
  verbose
)
}
\arguments{
\item{analyse_epochs}{the epochs to analyse, e.g. \code{c(1, 2, 3)},
as checked by \link{check_analyse_epochs}}

\item{args}{arguments for the \code{GCAE} executable}

\item{base_input_filename}{the base of the filenames that are
used as input for \code{GCAE}/\code{GCAE2}}

\item{base_output_filename}{the base of the filenames that are
used as output for \code{GCAE}/\code{GCAE2}}

\item{base_phenotype_value}{the base phenotypic value for an additive trait,
i.e. the phenotypic value for homozygotes of the common allele}

\item{bed_filename}{name of a PLINK genotypes (\code{.bed}) file,
as can be read using \link[plinkr]{read_plink_bed_file}}

\item{bim_filename}{name of a PLINK \code{.bim} file
as can be read using \link[plinkr]{read_plink_bim_file}}

\item{csv_filename}{name of the comma-seperated (\code{.csv}) file to be produced,
as checked by \link{check_csv_filename}}

\item{csv_filename_for_nmse}{name for the comma-seperated file
(as checked by \link{check_csv_filename})
to store the normalized mean square error compared to the identity line}

\item{data}{file prefix, not including path, of the data files.
The data files must be in EIGENSTRAT
or PLINK binary (\code{.bed}, \code{.bim}, \code{.fam}) format)}

\item{datadir}{directory where sample data is stored.
If not absolute: assumed relative to \verb{GenoCAE/} directory.
A valid \code{datadir} ends with a slash (\code{/})
and does not start with a full stop (\code{.}),
as to mimic the \code{GCAE} interface}

\item{data_opts_id}{data options id,
corresponding to a file data_opts/data_opts_id.json
The data options affect how data is input to the model.}

\item{epoch}{the epoch, as checked by \link{check_epoch}}

\item{epochs}{number of epochs, as checked by \link{check_epochs}}

\item{estimated_values}{estimated values}

\item{evaluate_filenames}{names of the files returned by
\link{gcae_evaluate}}

\item{evaluate_results}{the results of a call to
\link{parse_evaluate_filenames},
as checked by \link{check_evaluate_results}}

\item{example_filename}{name of the example file}

\item{fam_filename}{name of a PLINK \code{.fam} file
as can be read using \link[plinkr]{read_plink_fam_file}}

\item{file_extension}{a file extension, for example \code{.R}
or \code{.csv}.}

\item{folder_name}{name of a folder,
as checked by \link{check_folder_name}}

\item{gcae_experiment_params}{parameters to run a full
\code{GCAE} experimenr (i.e. call \link{do_gcae_experiment}),
as created by \link{create_gcae_experiment_params} and checked
by \link{check_gcae_experiment_params}}

\item{gcae_experiment_params_filename}{name of the file to
save/read \code{gcae_experiment_params} to/from,
using \link{read_gcae_experiment_params_file} and
\link{save_gcae_experiment_params}}

\item{gcae_experiment_results}{the results of a call
to \link{do_gcae_experiment},
as checked by \link{check_gcae_experiment_results}}

\item{gcae_folder}{folder where \code{GCAE} is installed}

\item{gcae_input_data}{in-memory data that can be used
as input for \code{GCAE}. Use link{read_gcae_input_files}
to read it from files, use \link{check_gcae_input_data}
to check it for validity.}

\item{gcae_input_filenames}{the GCAE input filenames,
in the form of a \link{list} with the following elements:
\itemize{
\item \code{bed_filename}: path to a \code{.bed} file,
as can be read by \link[plinkr]{read_plink_bed_file}
\item \code{bim_filename}: path to a \code{.bim} file,
as can be read by \link[plinkr]{read_plink_bim_file}
\item \code{fam_filename}: path to a \code{.fam} file,
as can be read by \link[plinkr]{read_plink_fam_file}
\item \code{phe_filename}: path to a \code{.phe} file,
as can be read by \link[plinkr]{read_plink_phe_file}
\item \code{labels_filename}: path to a \code{.csv} file
as can be read by \link{read_labels_file}
}}

\item{gcae_options}{options to run GCAE,
as created by \link{create_gcae_options}}

\item{gcae_options_filename}{name of a file to save/load a \code{gcae_options}
to/from}

\item{gcae_setup}{GCAE model setup,
as created by \link{create_gcae_setup},
read from file by \link{read_gcae_setup_file},
checked by \link{check_gcae_setup},
and saved to file by \link{save_gcae_setup}.}

\item{gcae_setup_filename}{name of a file to save/load a \code{gcae_setup}
to/from}

\item{gcae_version}{version of GCAE, e.g. \code{"1.0"}
Use \link{get_gcae_version} to get the \code{GCAE} version.}

\item{gcaer_folder}{name of the folder where \link{gcaer}
stores its temporary files}

\item{genotype_concordances_filename}{path to the
\code{genotype_concordances.csv} file}

\item{genotype_concordances_table}{genotype concordances
in time,
as can be checked by \link{check_genotype_concordances_table}}

\item{gif_filename}{name of a GIF file,
as checked by \link{check_gif_filename}}

\item{github_repo_branch_name}{name of the \code{git} branch,
e.g \code{master} or \code{Pheno}.
Use \link{get_default_github_repo_branch_name} for the default value}

\item{github_repo_url}{URL of the GitHub repository,
e.g. \verb{https://github.com/kausmees/GenoCAE}
or \verb{https://github.com/cnettel/GenoCAE}
or \verb{https://github.com/richelbilderbeek/GenoCAE}.
Use \link{get_default_github_repo_url} for the default value}

\item{labels_filename}{path to a \code{.csv} file containing super-
and -population labels, as can be read by \link{read_labels_file}}

\item{labels_table}{a table containing super-
and -population labels, as can be read by \link{read_labels_file}.
Note that \code{GCAE} does not use column names, \code{gcaer} adds these
for convenience}

\item{log_filename}{name of a \code{GCAE} \code{.log} file}

\item{losses_from_project_filename}{path to the
\code{losses_from_project.csv} file}

\item{losses_from_train_t_filename}{path to the
\code{losses_from_train_t.csv} file}

\item{losses_from_train_t_table}{losses from training, measured
in the training step,
as can be checked by \link{check_losses_from_train_t_table}}

\item{losses_from_train_v_filename}{path to the
\code{losses_from_train_v.csv} file}

\item{losses_from_train_v_table}{losses from training, measured
in the validation step,
as can be checked by \link{check_losses_from_train_v_table}}

\item{metric}{the metric, as checked by \link{check_metric}.
Use \code{""} to indicate not to use a clusting metric}

\item{metrics}{the metrics, as checked by \link{check_metrics}.
Use \code{""} to indicate not to use a clusting metric}

\item{model}{the \code{GCAE} architecture,
as checked by \link{check_model}.
Use \link{read_model_file} to read a \code{GCAE} architecture from file.
Use \link{save_model} to save the \code{model} to a JSON file.
Use \link{plot_model} to plot the \code{model}.
Use \link{save_model_plot} to save a plot of the \code{model} to file.}

\item{model_filename}{name of a file that stores a \code{GCAE} architecture,
as checked by \link{check_model_filename}.
Use \link{read_model_file} to read a \code{GCAE} architecture from that file.}

\item{model_id}{model id, corresponding to a file models/model_id.json}

\item{n_neurons}{the number of neurons,
as checked by \link{check_n_neurons}}

\item{nmse_in_time_filename}{name of the file for
the normalized mean squared error in time.}

\item{nmse_in_time_table}{normalized mean squared error in time,
as can be checked by \link{check_nmse_in_time_table}}

\item{ormr_folder_name}{folder where the link[ormr]{ormr} R package
installs the Python packages}

\item{os}{name of the operating system,
as returned by \link[rappdirs]{app_dir}}

\item{out}{the base filename of the output files.
This parameter is named after the \code{GCAE}
\code{--out} flag}

\item{phe_filename}{name of a PLINK phenotype (\code{.phe}) file
as can be read using \link[plinkr]{read_plink_phe_file}}

\item{plink_options}{PLINK options,
as created by \link[plinkr]{create_plink_options}}

\item{pheno_model_filename}{name of a file that stores
the phenotype prediction \code{GCAE} architecture,
as checked by \link{check_pheno_model_filename}.
Use \link{read_pheno_model_file} to read a \code{GCAE}
pehnotype prediction architecture from that file.}

\item{pheno_model_id}{phenotype model ID,
as checked by \link{check_pheno_model_id}}

\item{phenotype_predictions_filename}{name of a file for
phenotype predictions}

\item{phenotype_predictions_table}{a table of phenotype predictions,
with a true and estimated value for all individuals at a certain
epoch}

\item{png_filename}{name of the \code{PNG} file to be produced,
as checked by \link{check_png_filename}}

\item{project_filenames}{path to the files
as created (and returned) by \link{gcae_project}}

\item{python_bin_path}{path to a \code{python} binary/executable}

\item{python_version}{the Python version}

\item{resume_from}{the epoch to resume from,
as checked by \link{check_resume_from}}

\item{save_interval}{epoch intervals at which to save state of model,
and at which to calculate the valid loss}

\item{score_per_pop_filename}{name of a file containing the evaluation
scores from the dimensionality reduction}

\item{score_per_pop_table}{table that contains the dimensionality
reduction scores per population}

\item{scores_filename}{name of the files for F1 scores through time}

\item{scores_table}{table that contains the dimensionality
reduction scores}

\item{superpops}{path to the \code{superpops} file,
which contains the labels table and can be read
using \link{read_labels_file}.
This argument is named after the GCAE \code{--superpops} CLI flag.}

\item{train_filenames}{path to the files
as created (and returned) by \link{gcae_train}}

\item{train_times_table}{a table with training time per epoch,
as can be checked with \link{check_train_times_table}}

\item{train_opts_id}{train options id, corresponding to a file
train_opts/train_opts_id.json}

\item{train_times_filename}{path to the \code{train_times.csv} file}

\item{trainedmodeldir}{name of the folder to save the trained model to.
A valid \code{trainedmodeldir} ends with a slash (\code{/})
as to mimic the \code{GCAE} interface}

\item{trainedmodelname}{name of the trained model,
which is \code{ae_out} by default}

\item{true_values}{the true/known values}

\item{unique_id}{the unique identifier of a setting}

\item{url}{\code{GCAE} download URL}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link[plinkr]{check_verbose} to detect if this argument is valid.}
}
\description{
This function does nothing. It is intended to inherit is parameters'
documentation.
}
\note{
This is an internal function, so it should be marked with
\code{@noRd}. This is not done, as this will disallow all
functions to find the documentation parameters
}
\author{
Richèl J.C. Bilderbeek
}
