% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_gcae_script.R
\name{fix_gcae_script}
\alias{fix_gcae_script}
\title{Fix the GCAE script}
\usage{
fix_gcae_script(gcae_options)
}
\description{
Fix the GCAE script.
}
\details{
Since \verb{https://github.com/cnettel/GenoCAE/commit/}
\code{ed5c21c493b50e6da159b4ed4a70a454824f883f} the gcaer build broke,
due to the addition of:

\code{
  tf.config.experimental.enable_tensor_float_32_execution(
    False
  )
}

This function comments out that line.
}
