% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcae_evaluate.R
\name{gcae_evaluate}
\alias{gcae_evaluate}
\title{Use \code{GCAE} its evaluate}
\usage{
gcae_evaluate(gcae_setup, gcae_options, metrics, epoch, verbose = FALSE)
}
\description{
Use \code{GCAE} its evaluate
}
\author{
Richèl J.C. Bilderbeek
}
