% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcae_plot.R
\name{gcae_plot}
\alias{gcae_plot}
\title{Plot}
\usage{
gcae_plot(
  datadir,
  data,
  superpops = file.path(datadir, "HO_superpopulations"),
  gcae_setup = create_gcae_setup(),
  gcae_options = create_gcae_options(),
  verbose = FALSE
)
}
\arguments{
\item{datadir}{directory where sample data is stored.
If not absolute: assumed relative to \verb{GenoCAE/} directory.
A valid \code{datadir} ends with a slash (\code{/})
and does not start with a full stop (\code{.})}

\item{data}{file prefix, not including path, of the data files.
The data files must be in EIGENSTRAT
or PLINK binary (\code{.bed}, \code{.bim}, \code{.fam}) format)}

\item{superpops}{path to the superpops file.
This argument is named after the GCAE \code{--superpops} CLI flag.}

\item{gcae_setup}{GCAE model setup,
as created by \link{create_gcae_setup}}

\item{gcae_options}{options to run GCAE,
as created by \link{create_gcae_options}}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link{check_verbose} to detect if this argument is valid.}
}
\value{
full paths to the created files
}
\description{
Plot, using the GCAE command-line interface
}
\author{
Richèl J.C. Bilderbeek
}
