% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcae_train.R
\name{gcae_train}
\alias{gcae_train}
\title{Train model, and save its state at certain epochs to disk.}
\usage{
gcae_train(
  datadir = "example_tiny/",
  data = "HumanOrigins249_tiny",
  gcae_setup = create_gcae_setup(),
  epochs = 1,
  save_interval = 1,
  gcae_options = create_gcae_options(),
  verbose = FALSE
)
}
\arguments{
\item{datadir}{directory where sample data is stored}

\item{data}{file prefix, not including path, of the data files
(EIGENSTRAT of PLINK format)}

\item{epochs}{number of epochs to train}

\item{save_interval}{epoch intervals at which to save state of model,
and at which to calculate the valid loss}

\item{gcae_options}{options to run GCAE,
as created by \link{create_gcae_options}}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link{check_verbose} to detect if this argument is valid.}
}
\value{
full paths to the created files
}
\description{
Train model, and save its state at certain epochs to disk,
use the GCAE command-line interface
}
\author{
Richèl J.C. Bilderbeek
}
