% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcae_train.R
\name{gcae_train}
\alias{gcae_train}
\title{Train model, and save its state at certain epochs to disk.}
\usage{
gcae_train(
  gcae_setup = create_gcae_setup(),
  epochs = 1,
  save_interval = 1,
  gcae_options = create_gcae_options(),
  verbose = FALSE
)
}
\arguments{
\item{gcae_setup}{GCAE model setup,
as created by \link{create_gcae_setup}}

\item{epochs}{number of epochs to train}

\item{save_interval}{epoch intervals at which to save state of model,
and at which to calculate the valid loss}

\item{gcae_options}{options to run GCAE,
as created by \link{create_gcae_options}}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link[plinkr]{check_verbose} to detect if this argument is valid.}
}
\value{
full paths to the created files
}
\description{
Train model, and save its state at certain epochs to disk,
use the GCAE command-line interface
}
\examples{
if (plinkr::is_on_ci() && is_gcae_installed()) {
  gcae_setup <- create_gcae_setup(
    datadir = file.path(gcae_options$gcae_folder, "example_tiny/"),
    data = "issue_6_bin",
    model_id = "M1",
    pheno_model_id = "p2"
  )
  train_filenames <- gcae_train(
    gcae_setup = gcae_setup,
    epochs = 1,
    save_interval = 1
  )
}
}
\seealso{
use \link{create_gcae_train_args} to get the command-line
arguments
}
\author{
Richèl J.C. Bilderbeek
}
