% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gcae_data_opts_filename.R
\name{get_gcae_data_opts_filename}
\alias{get_gcae_data_opts_filename}
\title{Get the filename to the GCAE data options}
\usage{
get_gcae_data_opts_filename(
  data_opts_id = "b_0_4",
  gcae_options = create_gcae_options()
)
}
\arguments{
\item{data_opts_id}{data options id,
corresponding to a file data_opts/data_opts_id.json
The data options affect how data is input to the model.}

\item{gcae_options}{options to run GCAE,
as created by \link{create_gcae_options}}
}
\value{
full path to the data options' JSON file
}
\description{
Get the filename to the GCAE data options.
This function is named after the GCAE \code{data_opts} folder.
}
\examples{
get_gcae_data_opts_filename(
  data_opts_id = "b_0_4"
)
}
\author{
Richèl J.C. Bilderbeek
}
