% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gcae_help_text.R
\name{get_gcae_help_text}
\alias{get_gcae_help_text}
\title{Get the text of the gcae help file}
\usage{
get_gcae_help_text(gcae_options = create_gcae_options())
}
\arguments{
\item{gcae_options}{options to run GCAE,
as created by \link{create_gcae_options}}
}
\value{
the text of the gcae help file
}
\description{
Get the text of the gcae help file
}
\examples{
if (is_gcae_installed()) {

  # Default version
  get_gcae_help_text()
}
}
\author{
Richèl J.C. Bilderbeek
}
