% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gcae_output_subfolder.R
\name{get_gcae_output_subfolder}
\alias{get_gcae_output_subfolder}
\title{Get the path to the GCAE output subfolder}
\usage{
get_gcae_output_subfolder(gcae_setup)
}
\arguments{
\item{gcae_setup}{GCAE model setup,
as created by \link{create_gcae_setup},
read from file by \link{read_gcae_setup_file},
checked by \link{check_gcae_setup},
and saved to file by \link{save_gcae_setup}.}
}
\value{
full paths to the created files
}
\description{
Get the path to the GCAE output subfolder.
This subfolder contains the output of one model
}
\examples{
gcae_setup <- create_test_gcae_setup()
get_gcae_output_subfolder(gcae_setup = gcae_setup)
}
\author{
Richèl J.C. Bilderbeek
}
