% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gcae_output_subfolder.R
\name{get_gcae_output_subfolder}
\alias{get_gcae_output_subfolder}
\title{Get the path to the GCAE output subfolder}
\usage{
get_gcae_output_subfolder(
  data,
  gcae_setup,
  gcae_options = create_gcae_options()
)
}
\arguments{
\item{data}{file prefix, not including path, of the data files
(EIGENSTRAT of PLINK format)}

\item{gcae_setup}{GCAE model setup,
as created by \link{create_gcae_setup}}

\item{gcae_options}{options to run GCAE,
as created by \link{create_gcae_options}}
}
\value{
full paths to the created files
}
\description{
Get the path to the GCAE output subfolder.
This subfolder contains the output of one model
}
\examples{
get_gcae_output_subfolder(
  data = "HumanOrigins249_tiny",
  gcae_setup = create_gcae_setup()
)
}
\author{
Richèl J.C. Bilderbeek
}
