% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gcae_trainedmodelname.R
\name{get_gcae_trainedmodelname}
\alias{get_gcae_trainedmodelname}
\title{Get the GCAE \code{trainedmodelname}}
\usage{
get_gcae_trainedmodelname(data, gcae_setup)
}
\arguments{
\item{data}{file prefix, not including path, of the data files
(EIGENSTRAT of PLINK format)}

\item{gcae_setup}{GCAE model setup,
as created by \link{create_gcae_setup}}
}
\value{
the GCAE trained model name
}
\description{
Get the GCAE trained model name,
to be used by the GCAE\code{--trainedmodelname} CLI flag
}
\examples{
get_gcae_trainedmodelname(
  data = "HumanOrigins249_tiny",
  gcae_setup = create_gcae_setup()
)
}
\author{
Richèl J.C. Bilderbeek
}
