% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_n_neurons_in_latent_layer.R
\name{get_n_neurons_in_latent_layer}
\alias{get_n_neurons_in_latent_layer}
\title{Get the number of neurons in the layent layer}
\usage{
get_n_neurons_in_latent_layer(x)
}
\arguments{
\item{x}{an input of type \code{model} or \code{gcae_experiment_params}}
}
\value{
the number of neurons in the latent layer
}
\description{
Get the number of neurons in the layent layer
}
\examples{

# A model
if (is_gcae_repo_cloned()) {
  # A real GCAE file
  model_filename <- get_gcae_model_filename("M1")
} else {
  # An example file
  model_filename <- get_gcaer_filename("M0.json")
}
model <- read_model_file(model_filename)
get_n_neurons_in_latent_layer(model)
}
\author{
Richèl J.C. Bilderbeek
}
