% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_gcae.R
\name{install_gcae}
\alias{install_gcae}
\title{Install a specific version of GCAE}
\usage{
install_gcae(
  gcae_options = create_gcae_options(),
  github_repo_url = get_default_github_repo_url(),
  github_repo_branch_name = get_default_github_repo_branch_name(),
  verbose = FALSE
)
}
\arguments{
\item{gcae_options}{options to run GCAE,
as created by \link{create_gcae_options}}

\item{github_repo_url}{URL of the GitHub repository,
e.g. \verb{https://github.com/kausmees/GenoCAE}
or \verb{https://github.com/cnettel/GenoCAE}
or \verb{https://github.com/richelbilderbeek/GenoCAE}.
Use \link{get_default_github_repo_url} for the default value}

\item{github_repo_branch_name}{name of the \code{git} branch,
e.g \code{master} or \code{Pheno}.
Use \link{get_default_github_repo_branch_name} for the default value}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link[plinkr]{check_verbose} to detect if this argument is valid.}
}
\value{
Nothing
}
\description{
Install a specific version of GCAE.
}
\details{
\enumerate{
\item Clone the GCAE repo, see \link{clone_gcae_repo}
\item Install the needed Python packages, see \link{install_gcae_requirements}
\item Fix the GCAE script, see \link{fix_gcae_script}
}
}
\author{
Richèl J.C. Bilderbeek
}
