% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_gcae_installed.R
\name{is_gcae_installed}
\alias{is_gcae_installed}
\title{Determine if \code{GCAE} is installed}
\usage{
is_gcae_installed(gcae_options = create_gcae_options())
}
\arguments{
\item{gcae_options}{options to run GCAE,
as created by \link{create_gcae_options}}
}
\value{
\code{TRUE} if \code{GCAE} is installed
}
\description{
Determine if \code{GCAE} is installed
}
\examples{
is_gcae_installed()
}
\author{
Richèl J.C. Bilderbeek
}
