% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_gcae_installed.R
\name{is_gcae_installed}
\alias{is_gcae_installed}
\title{Determine if \code{GCAE} is installed}
\usage{
is_gcae_installed(gcae_options = create_gcae_options(), verbose = FALSE)
}
\arguments{
\item{gcae_options}{options to run GCAE,
as created by \link{create_gcae_options}}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link[plinkr]{check_verbose} to detect if this argument is valid.}
}
\value{
\code{TRUE} if \code{GCAE} is installed
}
\description{
Determine if \code{GCAE} is installed
}
\note{
this is a slow function.

A quicker, yet less accurate way to determine if \code{GCAE} is
installed, is \link{has_cloned_gcae_repo}
}
\examples{
# Tip: 'has_cloned_gcae_repo' is a a quicker, yet less accurate way
# to determine if `GCAE` is installed
is_gcae_installed()
}
\author{
Richèl J.C. Bilderbeek
}
