% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_model_file.R
\name{read_model_file}
\alias{read_model_file}
\title{Read a \code{GCAE} labels file}
\usage{
read_model_file(model_filename)
}
\value{
a \code{labels_table}, as can be checked by \link{check_labels_table}
}
\description{
Read a \code{GCAE} labels file
}
\note{
\code{GCAE} does not save this data with column names.
}
\examples{
labels_filename <- get_gcaer_filename("gcae_input_files_1_labels.csv")
read_labels_file(labels_filename)
}
\seealso{
use \link{save_labels_table} to save a \code{labels_table}
}
\author{
Richèl J.C. Bilderbeek
}
