% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_files.R
\name{rename_files}
\alias{rename_files}
\title{Rename multiple files}
\usage{
rename_files(filenames_from, filenames_to)
}
\arguments{
\item{filenames_from}{one or more filenames}

\item{filenames_to}{one or more filenames}
}
\value{
nothing
}
\description{
Rename multiple files
}
\examples{
# Rename one file, same as file.rename
tempfile_1 <- get_gcaer_tempfilename()
tempfile_2 <- get_gcaer_tempfilename()

file.create(tempfile_1)

rename_files(
  filenames_from = tempfile_1,
  filenames_to = tempfile_2
)

# Cleanup
file.remove(tempfile_2)

# Rename two files
tempfile_1 <- get_gcaer_tempfilename()
tempfile_2 <- get_gcaer_tempfilename()
tempfile_3 <- get_gcaer_tempfilename()
tempfile_4 <- get_gcaer_tempfilename()

file.create(tempfile_1)
file.create(tempfile_2)

rename_files(
  filenames_from = c(tempfile_1, tempfile_2),
  filenames_to = c(tempfile_3, tempfile_4)
)

# Cleanup
file.remove(tempfile_3)
file.remove(tempfile_4)
}
\author{
Richèl J.C. Bilderbeek
}
