% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resize_to_shared_individuals_from_data.R
\name{resize_to_shared_individuals_from_data}
\alias{resize_to_shared_individuals_from_data}
\title{Resize the \code{gcae_input_data} to only contain the shared individuals.}
\usage{
resize_to_shared_individuals_from_data(gcae_input_data, verbose = FALSE)
}
\arguments{
\item{gcae_input_data}{in-memory data that can be used
as input for \code{GCAE}. Use link{read_gcae_input_files}
to read it from files, use \link{check_gcae_input_data}
to check it for validity.}

\item{verbose}{the verbosity of a function.
Set to \link{TRUE} for more output.
Use \link[plinkr]{check_verbose} to detect if this argument is valid.}
}
\value{
the \code{gcae_input_data} with only shared individuals
}
\description{
Resize the \code{gcae_input_data} to only contain the shared individuals.
}
\examples{
gcae_input_data <- create_test_gcae_input_data()

# Show before
summarise_gcae_input_data(gcae_input_data)

gcae_input_data <- resize_to_shared_individuals_from_data(gcae_input_data)

# Show after
summarise_gcae_input_data(gcae_input_data)
}
\seealso{
Related resize functions:
\itemize{
\item Use \link{resize_to_shared_individuals_from_data}
to resize the data in in-memory data
\item Use \link{resize_to_shared_individuals_from_files}
to resize the data in files
}
}
\author{
Richèl J.C. Bilderbeek
}
