% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_gcae_experiment_results.R
\name{save_gcae_experiment_results}
\alias{save_gcae_experiment_results}
\title{Save a \code{gcae_experiment_results} to file}
\usage{
save_gcae_experiment_results(gcae_experiment_results, folder_name)
}
\arguments{
\item{gcae_experiment_results}{the results of a call
to \link{do_gcae_experiment},
as checked by \link{check_gcae_experiment_results}}
}
\value{
Nothing
}
\description{
Save a \code{gcae_experiment_results} to file
}
\examples{
# Create a gcae_experiment_results
gcae_experiment_results <- create_test_gcae_experiment_results()

# Use a temporary file
folder_name <- get_gcaer_tempfilename()

# Save gcae_experiment_results to temporary file
filenames <- save_gcae_experiment_results(
  gcae_experiment_results = gcae_experiment_results,
 folder_name = folder_name
)

# Cleanup
file.remove(unlist(filenames))
}
\author{
Richèl J.C. Bilderbeek
}
