% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_n_neurons_in_latent_layer.R
\name{set_n_neurons_in_latent_layer}
\alias{set_n_neurons_in_latent_layer}
\title{Set the number of neurons in the latent layer}
\usage{
set_n_neurons_in_latent_layer(model, n_neurons)
}
\value{
the model with the desired number of neurons in the latent layer
}
\description{
Set the number of neurons in the latent layer
}
\examples{
model_filename <- get_gcae_model_filename("M1")
model <- read_model_file(model_filename)
model <- set_n_neurons_in_latent_layer(
  model = model,
  n_neurons = 1
)
}
\author{
Richèl J.C. Bilderbeek
}
